% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catchRate}
\alias{catchRate}
\title{Catch Rate Dataset}
\format{
A data frame with 5 variables:
\describe{
\item{Ndex}{\code{integer} National Pokédex number}
\item{name}{\code{character} Name of the Pokémon}
\item{catchRate}{\code{integer} Catch rate of the Pokémon}
\item{note}{\code{character} Additional notes about the catch rate}
\item{Hdex}{\code{integer} Pokémon Legends: Arceus Pokédex number}
}
}
\source{
Bulbapedia
}
\usage{
catchRate
}
\description{
This dataset provides the catch rate for each Pokémon. The catch rate is a
statistic that determines how easy it is to catch a Pokémon. The higher the
catch rate, the easier it is to catch the Pokémon. In the context of Pokémon Legends: Arceus,
many Pokémon were given heightened catch rates to make catching them outside of battle much easier.
}
\keyword{datasets}
