% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\name{check_h}
\alias{check_h}
\title{Checks if a given xml_node type of tag exists}
\usage{
check_h(x, tag)
}
\arguments{
\item{x}{The object to search}

\item{tag}{What xml_node to search for}
}
\value{
TRUE if it exists FALSE otherwise
}
\description{
Checks if a given xml_node type of tag exists
}
