% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{damage_base}
\alias{damage_base}
\title{Pokemon Damage Calculation Base}
\usage{
damage_base(level, power, attackStat, defenseStat, .round = floor)
}
\arguments{
\item{level}{The level of the attacking Pokemon.}

\item{power}{The power of the used move.}

\item{attackStat}{The effective Attack stat of the attacking Pokemon.}

\item{defenseStat}{The effective Defense stat of the target Pokemon.}

\item{.round}{The rounding function to use. Gen I-IV uses \code{floor()} (truncated) and Gen V+ uses \code{round()} (.5 and under is truncated, above is rounded up).}
}
\value{
The base damage.
}
\description{
This function calculates the base damage inflicted during a Pokemon battle.
Gen I-IV uses \code{.round = floor}, while Gen V+ uses \code{.round = round}
}
