% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{damage_random}
\alias{damage_random}
\title{Pokemon Damage Calculation Random}
\usage{
damage_random(random, .min = 85, .max = 100)
}
\arguments{
\item{random}{How to generate the random number.}

\item{.min}{The min random number to generate to be divided by \code{.max}.}

\item{.max}{The max random number to generate to be divided by \code{.max}.}
}
\value{
The random factor.
}
\description{
This function calculates the random factor in Pokemon damage calculation.
Gen I-II uses \verb{.min = 217, .max = 255}, while Gen III+ uses \verb{.min = 85, .max = 100}
}
