% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{.name_from_row}
\alias{.name_from_row}
\title{Renames columns of a dataframe based on a specific row}
\usage{
.name_from_row(
  df,
  row_number = 1,
  prepend = "",
  postpend = "",
  test = TRUE,
  bind = TRUE
)
}
\arguments{
\item{df}{A dataframe to operate on}

\item{row_number}{The row number to use for renaming columns, default is 1}

\item{prepend}{A string to prepend to the new column names, default is ""}

\item{postpend}{A string to append to the new column names, default is ""}

\item{test}{A logical value indicating whether to check for unique column names, default is TRUE}

\item{bind}{A logical value indicating whether to bind rows with duplicate values, default is TRUE}
}
\value{
A dataframe with renamed columns
}
\description{
Renames columns of a dataframe based on a specific row
}
\examples{
# Deal with the odd header
# A tibble: 152 × 6
# 	Ndex  MS    English    Japanese   Japanese    Japanese
# <chr> <chr> <chr>      <chr>      <chr>       <chr>
# 	1 Ndex  "MS"  English    Kana       Hepburn     Trademarked
# 2 #0001 ""    Bulbasaur  フシギダネ Fushigidane Fushigidane
# 3 #0002 ""    Ivysaur    フシギソウ Fushigisō   Fushigisou
# 4 #0003 ""    Venusaur   フシギバナ Fushigibana Fushigibana
}
