% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{drop_all_na}
\alias{drop_all_na}
\title{Drops all rows with NA values in specified columns}
\usage{
drop_all_na(df, ...)
}
\arguments{
\item{df}{A dataframe to operate on}

\item{...}{Columns to check for NA values. If no columns are specified, all columns are checked.}
}
\value{
A dataframe with rows containing NA values in the specified columns removed
}
\description{
Drops all rows with NA values in specified columns
}
