% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Drops all columns with NA values in a dataframe}
\usage{
drop_na_columns(df)
}
\arguments{
\item{df}{A dataframe to operate on}
}
\value{
A dataframe with columns containing NA values removed
}
\description{
Drops all columns with NA values in a dataframe
}
