% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{efect_fun}
\alias{efect_fun}
\title{Compute Effectiveness Function}
\usage{
efect_fun(
  eft,
  efect = effectiveness_list$Standard,
  stab = FALSE,
  stab_multiplyer = stab_multiplyer$Standard
)
}
\arguments{
\item{eft}{Numeric value representing the effectiveness of an attack.}

\item{efect}{Numeric vector representing the effectiveness list. Default is effectiveness_list$Standard.}

\item{stab}{Logical value indicating whether the attack is of the same type as the attacker. Default is FALSE.}

\item{stab_multiplyer}{Numeric value representing the STAB multiplier. Default is stab_multiplyer$Standard.}
}
\value{
Numeric value representing the computed effectiveness.
}
\description{
This function computes the effectiveness of an attack based on the effectiveness value and the effectiveness list.
It uses a logarithmic mapping to determine the effectiveness.
}
\examples{
efect_fun(2, effectiveness_list$Standard)
}
