% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{effectiveness}
\alias{effectiveness}
\title{Compute Effectiveness of an Attack}
\usage{
effectiveness(
  attack,
  defender,
  stab = FALSE,
  stab_mult = stab_multiplyer$Standard,
  efect = effectiveness_list$Standard
)
}
\arguments{
\item{attack}{Character string representing the attacking type.}

\item{defender}{Character vector representing the defending types.}

\item{stab}{Logical value indicating whether the attack is of the same type as the attacker. Default is FALSE.}

\item{stab_mult}{Numeric value representing the STAB multiplier. Default is stab_multiplyer$Standard.}

\item{efect}{Numeric vector representing the effectiveness list. Default is effectiveness_list$Standard.}
}
\value{
Numeric value representing the effectiveness of the attack.
}
\description{
This function computes the effectiveness of an attack against a defender.
It uses the type chart to determine the effectiveness of the attack.
}
\examples{
effectiveness("Fire", c("Grass", "Bug"), effectiveness_list$Standard)
}
