% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\docType{data}
\name{effectiveness_list}
\alias{effectiveness_list}
\title{List of Effectiveness Values}
\format{
list
}
\source{
Bulbapedia
}
\usage{
effectiveness_list
}
\description{
This list contains various sets of effectiveness values used in different Pokémon games.
Each set of values represents a different game or game mechanic.
}
\examples{
effectiveness_list$Standard
effectiveness_list$LegendsArceus
effectiveness_list$MysteryDungeonRedBlue
effectiveness_list$ExplorersTimeDarknessSky
effectiveness_list$ErraticIQskill
effectiveness_list$Rumble
effectiveness_list$Shuffle
effectiveness_list$GO
effectiveness_list$GO2018
effectiveness_list$GO2017
}
\keyword{datasets}
