% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{effectivenessPokemon}
\alias{effectivenessPokemon}
\title{Compute Effectiveness of an Attack for a Pokémon}
\usage{
effectivenessPokemon(
  attack,
  pokemon,
  regional = NA_character_,
  form = NA_character_,
  efect = effectiveness_list$Standard,
  stab = FALSE,
  stab_mult = stab_multiplyer$Standard
)
}
\arguments{
\item{attack}{Character string representing the attacking type.}

\item{pokemon}{Either a character string representing the Pokémon's name or a data frame containing Pokémon data.}

\item{regional}{Character string representing the regional form of the Pokémon. Default is NA.}

\item{form}{Character string representing the form of the Pokémon. Default is NA.}

\item{efect}{Numeric vector representing the effectiveness list. Default is effectiveness_list$Standard.}

\item{stab}{Logical value indicating whether the attack is of the same type as the attacker. Default is FALSE.}

\item{stab_mult}{Numeric value representing the STAB multiplier. Default is stab_multiplyer$Standard.}
}
\value{
Numeric value representing the effectiveness of the attack.
}
\description{
This function computes the effectiveness of an attack against a specific Pokémon.
It uses the type chart to determine the effectiveness of the attack.
}
\examples{
effectivenessPokemon("Fire", "Zekrom", efect = effectiveness_list$Standard)
}
