% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evYield}
\alias{evYield}
\title{Pokémon Effort Value Yields}
\format{
A data frame with 16 variables:
\describe{
\item{\code{ndex}}{\code{integer} National Pokedex number. This is the unique identifier for each Pokémon in the National Pokédex.}
\item{\code{name}}{\code{character} Name of the Pokemon. This is the common name used to identify a particular Pokémon.}
\item{\code{form}}{\code{character} Form of the Pokemon. This refers to the different physical forms that Pokémon can take. NA if the Pokemon does not have a special form.}
\item{\code{regional}}{\code{factor} Regional form of the Pokemon. This refers to the variations of Pokémon that appear differently and have different attributes compared to their counterparts in other regions. NA if the Pokemon does not have a regional form.}
\item{\code{Gen}}{\code{factor} The generation the Pokemon was introduced in. This refers to the different stages of Pokémon games, each introducing new Pokémon, regions, and gameplay concepts.}
\item{\code{Mega}}{\code{character} The Mega form. These are special forms that certain Pokémon can take.}
\item{\code{MegaOrPrimal}}{\code{factor} Indicates if the Pokémon is in a Mega or Primal form. These are special forms that certain Pokémon can take.}
\item{\code{Exp}}{\code{integer} The experience yield of the Pokemon.}
\item{\code{note}}{\code{character} Any additional notes about the Pokemon. NA if there are no additional notes.}
\item{\code{HP}}{\code{integer} The HP effort value yield of the Pokemon.}
\item{\code{Attack}}{\code{integer} The Attack effort value yield of the Pokemon.}
\item{\code{Defense}}{\code{integer} The Defense effort value yield of the Pokemon.}
\item{\code{SpAttack}}{\code{integer} The Special Attack effort value yield of the Pokemon.}
\item{\code{SpDefense}}{\code{integer} The Special Defense effort value yield of the Pokemon.}
\item{\code{Speed}}{\code{integer} The Speed effort value yield of the Pokemon.}
\item{\code{TotalEVs}}{\code{integer} The total effort value yield of the Pokemon.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_effort_value_yield_(Generation_III)}{List of Pokémon by effort value yield (Generation III). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_effort_value_yield_(Generation_IV)}{List of Pokémon by effort value yield (Generation IV). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_effort_value_yield_(Generation_VII)}{List of Pokémon by effort value yield (Generation VII). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_effort_value_yield_(Generation_VIII)}{List of Pokémon by effort value yield (Generation VIII). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_effort_value_yield_(Generation_IX)}{List of Pokémon by effort value yield (Generation IX). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_effort_value_yield_(Generation_V-VI)}{List of Pokémon by effort value yield (Generation V-VI). In Bulbapedia.}
}
\usage{
evYield
}
\description{
This dataset provides information about Pokémon effort value yields as per each generation,
which includes Pokémon from the Kanto region first, followed by those from Johto,
Hoenn, Sinnoh, Unova, Kalos, Alola, Galar, Hisui, and Paldea.
}
\details{
A dataset about Pokémon effort value yields per generation.
When a Pokémon is defeated in battle, it will give effort values to the Pokémon that participated in the battle against it. The values shown below are the effort points that a Pokémon will give in.
}
\examples{
data(evYield)
}
\references{
\href{https://bulbapedia.bulbagarden.net/wiki/Effort_values}{Effort Values. In Bulbapedia.}
}
\keyword{datasets}
