% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{factor_type}
\alias{factor_type}
\title{Convert Pokémon Types to Factor}
\usage{
factor_type(x)
}
\arguments{
\item{x}{A character vector of Pokémon types.}
}
\value{
A factor with levels defined by the unique Pokémon types.
}
\description{
This function converts a vector of Pokémon types to a factor with levels defined by the unique Pokémon types.
}
\examples{
# Assume 'types' is a dataframe with a column 'types' containing unique Pokémon types
pokemon_types <- c("Fire", "Water", "Grass", "Electric")
factor_types <- factor_type(pokemon_types)

}
