% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{family}
\alias{family}
\title{Pokemon Evolution Dataset}
\format{
A data frame with the following variables:
\describe{
\item{family}{\code{character} The family to which the Pokemon belongs.}
\item{base}{\code{character} The prior form of the pokemon or the base form}
\item{form}{\code{character} The specific form of the Pokemon. NA if not applicable.}
\item{method}{\code{character} The method of evolution.}
\item{evo}{\code{character} The next stage of the Pokemon.}
\item{evoForm}{\code{character} The form of next stage of the Pokemon. NA if not applicable.}
\item{transition}{\code{character} The transition stage of the Pokemon.}
\item{first}{\code{character} The first form of the Pokemon.}
\item{firstForm}{\code{character} The specific first form of the Pokemon. NA if not applicable.}
\item{regional}{\code{character} The regional form of the Pokemon. NA if not applicable.}
\item{evoRegional}{\code{character} The regional evolved form of the Pokemon. NA if not applicable.}
\item{firstRegional}{\code{character} The regional first form of the Pokemon. NA if not applicable.}
}
}
\source{
Bulbapedia
}
\usage{
family
}
\description{
This dataset contains evolution and family information.
}
\keyword{datasets}
