% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{familyLong}
\alias{familyLong}
\title{Pokemon Dataset}
\format{
A data frame with the following variables:
\describe{
\item{family.x}{The family of the Pokemon}
\item{transition}{The transition stage of the Pokemon}
\item{first}{The first form of the Pokemon}
\item{firstForm}{The first form of the Pokemon if it has multiple forms}
\item{firstRegional}{The first regional form of the Pokemon}
\item{name}{The name of the Pokemon}
\item{form}{The form of the Pokemon}
\item{regional}{The regional form of the Pokemon}
\item{stage}{The stage of the Pokemon}
\item{family.y}{The legendary family of the Pokemon}
\item{size}{The size of the legendary family}
}
}
\source{
Bulbapedia
}
\usage{
familyLong
}
\description{
This dataset contains evolution and family information.
}
\keyword{datasets}
