% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{fill_missing_ndex}
\alias{fill_missing_ndex}
\title{Fill Missing 'ndex' Values in a Dataset}
\usage{
fill_missing_ndex(df, known_df, name_var, ndex_var)
}
\arguments{
\item{df}{A tibble or data frame containing the main dataset.}

\item{known_df}{A tibble or data frame containing the known dataset.}

\item{name_var}{The name of the 'name' column in the datasets (as a symbol or a string).}

\item{ndex_var}{The name of the 'ndex' column in the datasets (as a symbol or a string).}
}
\value{
A tibble or data frame with the missing 'ndex' values filled.
}
\description{
This function fills missing 'ndex' values in a dataset based on the 'name' column.
It uses a known dataset to look up the missing 'ndex' values.
}
