% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{filter_non_unique}
\alias{filter_non_unique}
\title{Filters rows of a dataframe based on non-unique values in specified columns}
\usage{
filter_non_unique(df, ...)
}
\arguments{
\item{df}{A dataframe to operate on.}

\item{...}{Columns to check for non-unique values. If no columns are specified, all columns are checked.}
}
\value{
A dataframe with rows containing non-unique values in the specified columns removed.
}
\description{
This function filters rows of a dataframe based on non-unique values in specified columns. If no columns are specified, all columns are checked.
}
\examples{
library(tibble)
filter_non_unique(
  tribble(
    ~a, ~b,
    1, 3,
    2, 4,
    2, 4
  )
)
}
