% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{fun}
\alias{fun}
\title{Match a string over a regex vector}
\usage{
fun(names_vector, match, val)
}
\arguments{
\item{match}{What to check for a vector RegEx}

\item{val}{The value to return}

\item{name}{The value(s) to test}
}
\description{
takes a vector of names, matches them against patterns using str_detect,
and then returns the corresponding values from patternsT[\link{val}]
where the match is TRUE.
}
