% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{gen_catch}
\alias{gen_catch}
\title{Generate Catch Rate Pokémon Data}
\usage{
gen_catch(write = FALSE, root = "data/", file = "PokemonCatch")
}
\arguments{
\item{write}{Logical. If TRUE, the function will write the data to a file. Default is FALSE.}

\item{root}{Character. The root directory where the file will be saved. Default is "data/".}

\item{file}{Character. The name of the file where the data will be saved. Default is "PokemonCatch".}
}
\value{
A tibble of the scraped Pokémon data.
}
\description{
This function scrapes data from Bulbapedia's "List of Pokémon by catch rate" page and returns a tibble of the data.
It also has the option to write the data to a file.
}
