% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{gen_evos}
\alias{gen_evos}
\title{Generate Pokémon Evolution Data}
\usage{
gen_evos(
  write = FALSE,
  root = "data/",
  file = "PokemonFamily",
  fileLong = "PokemonFamilyLong"
)
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the data frame to a csv file. Default is \code{FALSE}.}

\item{root}{The root directory where the csv file will be written if \code{write = TRUE}. Default is "data/".}

\item{file}{The name of the csv file to be written if \code{write = TRUE}. Default is "PokemonFamily".}

\item{fileLong}{The name of the long format csv file to be written if \code{write = TRUE}. Default is "PokemonFamilyLong".}
}
\value{
A data frame of Pokémon evolution data.
}
\description{
This function scrapes Pokémon evolution data from Bulbapedia and returns a data frame. If \code{write = TRUE}, it also writes the data frame to a csv file.
}
