% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVYield.R
\name{gen_evYield}
\alias{gen_evYield}
\title{Generate EV Yield}
\usage{
gen_evYield(write = FALSE, root = "data/", file = "PokemonEVYield")
}
\arguments{
\item{write}{A logical value indicating whether to write the output dataframe to a file. Default is FALSE.}

\item{root}{A character string specifying the root directory where the output file will be saved. Default is "data/".}

\item{file}{A character string specifying the name of the output file. Default is "PokemonEVYield".}
}
\value{
A dataframe of Pokémon effort value yields by generation.
}
\description{
This function scrapes data from Bulbapedia to generate a dataframe of Pokémon effort value yields by generation.
}
