% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{gen_groups}
\alias{gen_groups}
\title{Generate a tibble of legendary/mythical groups of Pokemon}
\usage{
gen_groups(write = FALSE, root = "data/", file = "PokemonGroups")
}
\arguments{
\item{write}{Logical, if TRUE the data is saved to a file, default is FALSE.}

\item{root}{Character, the root directory where the data file will be saved, default is "data/".}

\item{file}{Character, the name of the file where the data will be saved, default is "PokemonGroups".}
}
\value{
A tibble of legendary/mythical groups of Pokemon.
}
\description{
This function scrapes data from Bulbapedia to create a tibble of legendary/mythical groups of Pokemon.
}
