% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight.R
\name{gen_height}
\alias{gen_height}
\title{Generate Pokemon Height Data}
\usage{
gen_height(write = FALSE, root = "data/", file = "PokemonHeight")
}
\arguments{
\item{write}{A logical value indicating whether to write the data to a file.}

\item{root}{The root directory where the file should be saved.}

\item{file}{The name of the file where the data should be saved.}
}
\value{
A tibble containing the Pokemon height data.
}
\description{
Function to scrape and generate Pokemon height data from Bulbapedia.
}
