% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lang.R
\name{gen_lang}
\alias{gen_lang}
\title{Generate the National Dex with Other Language Data}
\usage{
gen_lang(write = FALSE, root = "data/", file = "PokemonLang")
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the tibble to a csv file. Default is \code{FALSE}.}

\item{root}{The root directory where the csv file will be written if \code{write = TRUE}. Default is "data/".}

\item{file}{The name of the csv file to be written if \code{write = TRUE}. Default is "PokemonLang".}
}
\value{
A tibble of the National Dex with language information.
}
\description{
This function generates a tibble of the National Dex with language information.
}
