% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mega.R
\name{gen_mega}
\alias{gen_mega}
\title{Generate Mega Evolution and Primal Reversion Data}
\usage{
gen_mega(write = FALSE, root = "data/", file = "PokemonMega")
}
\arguments{
\item{write}{Logical, if TRUE, the function writes the data to a CSV file. Default is FALSE.}

\item{root}{Character, the root directory where the CSV file will be saved. Default is "data/".}

\item{file}{Character, the name of the CSV file. Default is "PokemonMega".}
}
\value{
A dataframe containing the Mega Evolution and Primal Reversion data.
}
\description{
This function scrapes data from Bulbapedia about Mega Evolutions and Primal Reversions in Pokémon.
It processes the data and optionally writes it to a CSV file.
}
