% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight.R
\name{gen_physicalAttr}
\alias{gen_physicalAttr}
\title{Generate Physical Attributes of Pokemon}
\usage{
gen_physicalAttr(
  write = FALSE,
  root = "data/",
  file = "PokemonPhysicalAttr",
  fileHeight = "PokemonHeight",
  fileWeight = "PokemonWeight"
)
}
\arguments{
\item{write}{Logical, whether to write the resulting data frame to a file.}

\item{root}{Character, the root directory where the data files are located.}

\item{file}{Character, the name of the file to write the resulting data frame to.}

\item{fileHeight}{Character, the name of the file containing the height data.}

\item{fileWeight}{Character, the name of the file containing the weight data.}
}
\value{
A data frame of physical attributes of Pokemon.
}
\description{
This function generates a data frame of physical attributes (weight and height) of Pokemon.
It performs two inner joins on the weight and height data frames and combines the results.
The function can optionally write the resulting data frame to a file.
}
