% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regional.R
\name{gen_reginal}
\alias{gen_reginal}
\title{Generates regional dex from scratch}
\usage{
gen_reginal(write = FALSE)
}
\arguments{
\item{write}{Write the file to \code{data/PokemonRegional.csv}?}
}
\value{
A tibble with of regional dexes
}
\description{
Generates regional dex from scratch
}
