% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{gen_type}
\alias{gen_type}
\title{Generates Pokémon Types from Scratch}
\usage{
gen_type(write = FALSE, root = "data/", file = "PokemonTypes")
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the tibble to a CSV file. Default is \code{FALSE}.}

\item{root}{The root directory where the CSV file will be written if \code{write = TRUE}. Default is \code{"data/"}.}

\item{file}{The name of the CSV file to be written if \code{write = TRUE}. Default is \code{"PokemonTypes"}.}
}
\value{
A tibble with a single column named 'types' containing unique Pokémon types.
}
\description{
This function generates a tibble of Pokémon types by scraping the Bulbapedia website. It extracts the types from the HTML table on the page and creates a unique list of Pokémon types. It also adds two additional types, 'Stellar' and '???'. If \code{write = TRUE}, it writes the tibble to a CSV file.
}
