% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{gen_type_chart}
\alias{gen_type_chart}
\title{Generate Pokémon Type Chart}
\usage{
gen_type_chart(
  write = FALSE,
  root = "data/",
  file = "PokemonTypeChart",
  fileWide = "PokemonTypeChartWide"
)
}
\arguments{
\item{write}{Logical value indicating whether to write the generated charts to files. Default is FALSE.}

\item{root}{Character string specifying the root directory where the files should be written. Default is "data/".}

\item{file}{Character string specifying the filename for the long version of the chart. Default is "PokemonTypeChart".}

\item{fileWide}{Character string specifying the filename for the wide version of the chart. Default is "PokemonTypeChartWide".}
}
\value{
A list containing two data frames: 'typeChart' and 'typeChartWide'.
}
\description{
This function generates a Pokémon type chart by scraping data from Bulbapedia.
It creates two versions of the chart: a wide version and a long version.
The wide version has defending types as columns and attacking types as rows.
The long version is a reshaped version of the wide chart with three columns: attacking type, defending type, and effectiveness.
}
