% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight.R
\name{gen_weight}
\alias{gen_weight}
\title{Generate Pokemon Weight Data}
\usage{
gen_weight(write = FALSE, root = "data/", file = "PokemonWeight")
}
\arguments{
\item{write}{A logical value indicating whether to write the data to a file.}

\item{root}{The root directory where the file should be saved.}

\item{file}{The name of the file where the data should be saved.}
}
\value{
A tibble containing the Pokemon weight data.
}
\description{
Function to scrape and generate Pokemon weight data from Bulbapedia.
}
