% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{ggtypes}
\alias{ggtypes}
\title{Create a ggplot2 fill scale for Pokémon types}
\usage{
ggtypes(types = type_colors$new, aesthetics = "color")
}
\arguments{
\item{types}{A named vector representing a color palette for Pokémon types. The names of the vector are the Pokémon types, and the values are the corresponding colors.}

\item{aesthetics}{The aesthetic mapping to use for the fill scale. Default is "color".}
}
\value{
A ggplot2 fill scale.
}
\description{
This function creates a ggplot2 fill scale for Pokémon types using a specified color palette.
}
\examples{
# Create a ggplot2 fill scale for Pokémon types
ggtypes(types = type_colors$old)

# Create a ggplot2 fill scale for Pokémon types from scratch
ggtypes(types = c("Fire" = "red", "Water" = "blue"), aesthetics = "fill")
}
