% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups}
\alias{groups}
\title{Legendary and Mythical Pokémon Groupings}
\format{
A data frame with 3 variables:
\describe{
\item{family}{\code{character} The family grouping of the Pokémon. Examples include 'Tao trio', 'Super-ancient Pokémon', 'Paradox duo', etc.}
\item{name}{\code{character} The name of the individual Pokémon within the family grouping.}
\item{size}{\code{character} The size category of the family grouping. Categories include 'duos', 'trios', 'quartets', and 'quintets'.}
}
}
\source{
Bulbapedia
}
\usage{
groups
}
\description{
A dataset containing Legendary and Mythical Pokémon categorized by their family groupings and sizes.
}
\examples{
data(groups)
}
\keyword{datasets}
