% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{height}
\alias{height}
\title{Height Dataset}
\format{
A data frame with 8 variables:
\describe{
\item{name}{\code{character} Name of the Pokémon}
\item{form}{\code{character} Specific form of the Pokémon}
\item{ft}{\code{numeric} Height of the Pokémon in feet}
\item{m}{\code{numeric} Height of the Pokémon in meters}
\item{inch}{\code{numeric} Height of the Pokémon in inches}
\item{smallest}{\code{integer} Ranking of the Pokémon from smallest to biggest}
\item{bigest}{\code{integer} Ranking of the Pokémon from biggest to smallest}
\item{regional}{\code{character} Regional form of the Pokémon, if any}
\item{MegaOrPrimal}{\code{character} Is the Pokémon in a Mega or Primal form}
}
}
\source{
Bulbapedia
}
\usage{
height
}
\description{
This dataset contains information about the height of different Pokémon forms.
The height is represented in different units such as feet (ft), meters (m), and inches (inch).
It also includes the ranking of each Pokémon form from smallest to biggest,
and information about regional forms and Mega or Primal forms.
}
\keyword{datasets}
