% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mega}
\alias{mega}
\title{Pokémon Mega Evolution and Primal Reversion}
\format{
A data frame with 9 variables:
\describe{
\item{\code{ndex}}{\code{integer} The National Pokédex number of the Pokémon.}
\item{\code{name}}{\code{character} The name of the Pokémon.}
\item{\code{typeBefore}}{\code{factor} The primary type of the Pokémon before Mega Evolution or Primal Reversion.}
\item{\code{type2Before}}{\code{factor} The secondary type of the Pokémon before Mega Evolution or Primal Reversion, if applicable.}
\item{\code{abilityBefore}}{\code{character} The ability of the Pokémon before Mega Evolution or Primal Reversion.}
\item{\code{typeAfter}}{\code{factor} The primary type of the Pokémon after Mega Evolution or Primal Reversion.}
\item{\code{type2After}}{\code{factor} The secondary type of the Pokémon after Mega Evolution or Primal Reversion, if applicable.}
\item{\code{abilityAfter}}{\code{character} The ability of the Pokémon after Mega Evolution or Primal Reversion.}
\item{\code{stone}}{\code{character} The Mega Stone or Orb required for the Pokémon to undergo Mega Evolution or Primal Reversion.  Or a note in case of Rayquaza}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/Mega_Evolution}{Mega Evolution. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/Primal_Reversion}{Primal Reversion. In Bulbapedia.}
}
\usage{
mega
}
\description{
This dataset contains information about Pokémon that can undergo Mega Evolution and Primal Reversion,
which are special forms of evolution that Pokémon can undergo during battle to become more powerful.
These Pokémon are categorized by their National Pokédex number, name, type before and after evolution, ability before and after evolution, and the stone required for the evolution.
}
\details{
For example, ‘Rayquaza’ is a Dragon/Flying type Pokémon with the ability ‘Air Lock’ that can undergo
Mega Evolution to remain a Dragon/Flying type but with the ability ‘Delta Stream’. Unlike other Pokémon,
Rayquaza does not require a Mega Stone to undergo Mega Evolution. Instead, Rayquaza must swallow the
Meteorite and learn Dragon Ascent during the Delta Episode or must be taught Dragon Ascent via move tutor.

The 'Primal Reversion' refers to Kyogre and Groudon,
which are ancient Pokémon that revert to their primal forms with the Blue Orb and Red Orb respectively,
changing their abilities to 'Primordial Sea' and 'Desolate Land'.

Note that there exist Primal Dialga, hovever this appears to be different from Primal Reversion.
}
\examples{
data(mega)
}
\keyword{datasets}
