% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nationalDex}
\alias{nationalDex}
\title{Pokemon Data}
\format{
A data frame with 12 variables:
\describe{
\item{ndex}{National Pokedex number. Integer.}
\item{name}{Name of the Pokemon. Character.}
\item{form}{Form of the Pokemon. Character. NA if the Pokemon does not have a special form.}
\item{regional}{Regional form of the Pokemon. Character. NA if the Pokemon does not have a regional form.}
\item{type}{Primary type of the Pokemon. Character.}
\item{type2}{Secondary type of the Pokemon. Character. NA if the Pokemon does not have a secondary type.}
\item{generation}{The generation the Pokemon was introduced in. Integer.}
\item{isLegendary}{Whether the Pokemon is Legendary. Logical.}
\item{isMythical}{Whether the Pokemon is Mythical. Logical.}
\item{isUltraBeast}{Whether the Pokemon is an Ultra Beast. Logical.}
\item{family}{Legendary / Mythical family group of the Pokemon. Character. NA if the Pokemon does not belong to a family group.}
\item{size}{Legendary / Mythical family size. NA if the Pokemon does not belong to a family group.}
}
}
\source{
Bulbapedia
}
\usage{
nationalDex
}
\description{
Pokemon National dex information
}
\examples{
data(nationalDex)
}
\keyword{datasets}
