% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{non_unique}
\alias{non_unique}
\title{Checks for non-unique values in a vector}
\usage{
non_unique(x)
}
\arguments{
\item{x}{A vector to check for non-unique values}
}
\value{
A logical vector indicating which values are non-unique
}
\description{
Checks for non-unique values in a vector
}
