% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physicalAttr}
\alias{physicalAttr}
\title{Physical Attributes of Pokémon}
\format{
A data frame with 20 variables:
\describe{
\item{ndex}{\code{integer} National Pokédex number}
\item{name}{\code{character} Name of the Pokémon}
\item{regional}{\code{character} Regional variant, if any}
\item{type}{\code{character} Primary type of the Pokémon}
\item{type2}{\code{character} Secondary type of the Pokémon, if any}
\item{generation}{\code{integer} Generation of the Pokémon}
\item{isLegendary}{\code{logical} Indicates if the Pokémon is Legendary}
\item{isMythical}{\code{logical} Indicates if the Pokémon is Mythical}
\item{isUltraBeast}{\code{logical} Indicates if the Pokémon is an Ultra Beast}
\item{family}{\code{character} Family of the Pokémon, if any}
\item{size}{\code{character} Size of the Pokémon, if any}
\item{form}{\code{character} Form of the Pokémon, if any}
\item{lbs}{\code{numeric} Weight of the Pokémon in pounds}
\item{kg}{\code{numeric} Weight of the Pokémon in kilograms}
\item{lightest}{\code{integer} Lightest weight of the Pokémon in the Pokédex}
\item{heaveiest}{\code{integer} Heaviest weight of the Pokémon in the Pokédex}
\item{MegaOrPrimal}{\code{character} Is the Pokémon in a Mega or Primal form}
\item{ft}{\code{character} Height of the Pokémon in feet}
\item{m}{\code{numeric} Height of the Pokémon in meters}
\item{inch}{\code{integer} Height of the Pokémon in inches}
\item{smallest}{\code{integer} Smallest height of the Pokémon in the Pokédex}
\item{bigest}{\code{integer} Biggest height of the Pokémon in the Pokédex}
}
}
\source{
Bulbapedia
}
\usage{
physicalAttr
}
\description{
This dataset contains information about the physical attributes of Pokémon, including their type, generation, weight, height, and other characteristics.
}
\keyword{datasets}
