% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\name{pluck_matched_node}
\alias{pluck_matched_node}
\title{Extracts all xml_nodes matching the tag and all other elements that don't}
\usage{
pluck_matched_node(lst, tag)
}
\arguments{
\item{lst}{A list or xml_nodeset of xml_nodes to extract from}

\item{tag}{The tag to extract form the set}
}
\value{
A list of
matched_elements that have the matching xml_nodes
other_elements that have the updated list without the matched elements
}
\description{
Extracts all xml_nodes matching the tag and all other elements that don't
}
