% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{pokemon_damage}
\alias{pokemon_damage}
\title{Pokemon Damage Function}
\usage{
pokemon_damage(..., .gen = "V")
}
\arguments{
\item{.gen}{The generation to use for calculating damage. Default is "V".}
}
\value{
The calculated damage.
}
\description{
This function calculates the damage for a Pokemon based on the generation.
}
\seealso{
\code{\link{pokemon_damage_V}}, \code{\link{pokemon_damage_IV}}, \code{\link{pokemon_damage_III}}, \code{\link{pokemon_damage_II}}, \code{\link{pokemon_damage_I}} for the generation specific damage functions.
}
