% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{pokemon_damage_I}
\alias{pokemon_damage_I}
\title{Pokemon Damage Calculation Generation I}
\usage{
pokemon_damage_I(
  level,
  attackStat,
  defenseStat,
  power,
  effectiveness,
  stab = FALSE,
  critical = FALSE,
  random = "yes"
)
}
\arguments{
\item{level}{The level of the attacking Pokemon.}

\item{attackStat}{The effective Attack stat of the attacking Pokemon.}

\item{defenseStat}{The effective Defense stat of the target Pokemon.}

\item{power}{The power of the used move.}

\item{effectiveness}{The product of Type, and Type2.}

\item{stab}{Logical. If TRUE, add the same type attack bonus}

\item{critical}{Logical. If TRUE, it's a critical hit. Default is FALSE.}

\item{random}{Character string. If "yes", a random uniformly distributed integer between 217 and 255 is used. If "minmax", the min and max values are returned. If anything else, the average is used. Default is "yes".}
}
\value{
The calculated damage.
}
\description{
This function calculates the damage inflicted during a Pokemon battle.
It uses the formula from Generation I.
}
