% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{pokemon_damage_III}
\alias{pokemon_damage_III}
\title{Pokemon Damage Calculation Generation III}
\usage{
pokemon_damage_III(
  level,
  attackStat,
  defenseStat,
  power,
  effectiveness,
  burn = FALSE,
  screen = 1,
  targets = 1,
  weather = 1,
  FF = FALSE,
  stockpile = 1,
  critical = FALSE,
  doubleDmg = FALSE,
  charge = FALSE,
  HH = FALSE,
  stab = FALSE,
  random = "yes"
)
}
\arguments{
\item{level}{The level of the attacking Pokemon.}

\item{attackStat}{The effective Attack stat of the attacking Pokemon.}

\item{defenseStat}{The effective Defense stat of the target Pokemon.}

\item{power}{The power of the used move.}

\item{effectiveness}{The product of Type, and Type2.}

\item{burn}{Logical. If TRUE, the attacker is burned, its Ability is not Guts, and the used move is a physical move. Default is FALSE.}

\item{screen}{The screen multiplier. Default is 1.}

\item{targets}{The targets multiplier. Default is 1.}

\item{weather}{The weather multiplier. Default is 1.}

\item{FF}{Logical. If TRUE, the used move is Fire-type, and the attacker's Ability is Flash Fire that has been activated by a Fire-type move. Default is FALSE.}

\item{stockpile}{The stockpile multiplier. Default is 1.}

\item{critical}{Logical. If TRUE, it's a critical hit. Default is FALSE.}

\item{doubleDmg}{Logical. If TRUE, double the damage of the move. Default is FALSE.}

\item{charge}{Logical. If TRUE, the move is Electric-type and Charge takes effect. Default is FALSE.}

\item{HH}{Logical. If TRUE, the attacker's ally in a Double Battle has used Helping Hand on it. Default is FALSE.}

\item{stab}{Logical. If TRUE, add the same type attack bonus.}

\item{random}{Character string. If "yes", a random uniformly distributed integer between 85 and 100 is used. If "minmax", the min and max values are returned. If anything else, the average is used. Default is "yes".}
}
\value{
The calculated damage.
}
\description{
This function calculates the damage inflicted during a Pokemon battle.
It uses the formula from Generation III.
}
\note{
this function does not handle the specific conditions mentioned (like Reflect, Light Screen, Explosion, Selfdestruct, etc.).
You would need to handle these conditions before passing the values to this function
}
