% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{pokemon_damage_V}
\alias{pokemon_damage_V}
\title{Pokemon Damage Calculation Generation V and onward}
\usage{
pokemon_damage_V(
  level,
  attackStat,
  defenseStat,
  power,
  effectiveness,
  PB = FALSE,
  weather = 1,
  GlaiveRush = FALSE,
  critical = FALSE,
  stab = FALSE,
  burn = FALSE,
  other = 1,
  ZMove = FALSE,
  TeraShield = FALSE,
  targets = 1,
  random = "yes",
  .criticalVal = 1.5,
  .PBVal = 0.25
)
}
\arguments{
\item{level}{The level of the attacking Pokemon.}

\item{attackStat}{The effective Attack stat of the attacking Pokemon.}

\item{defenseStat}{The effective Defense stat of the target Pokemon.}

\item{power}{The power of the used move.}

\item{effectiveness}{The product of Type, and Type2.}

\item{PB}{Logical. If TRUE, the move is the second strike of Parental Bond. Default is FALSE.}

\item{weather}{The weather multiplier. Default is 1.}

\item{GlaiveRush}{Logical. If TRUE, the target used the move Glaive Rush in the previous turn. Default is FALSE.}

\item{critical}{Logical. If TRUE, it's a critical hit. Default is FALSE.}

\item{stab}{Logical. If TRUE, add the same type attack bonus.}

\item{burn}{Logical. If TRUE, the attacker is burned, its Ability is not Guts, and the used move is a physical move. Default is FALSE.}

\item{other}{The other multiplier. Default is 1.}

\item{ZMove}{Logical. If TRUE, the move is a Z-Move or Max Move and the target would be protected from that move (e.g. by Protect). Default is FALSE.}

\item{TeraShield}{Logical. If TRUE, it is applied in Tera Raid Battles when the Raid boss's shield is active. Default is FALSE.}

\item{targets}{The targets multiplier. Default is 1.}

\item{random}{Character string. If "yes", a random uniformly distributed integer between 85 and 100 is used. If "minmax", the min and max values are returned. If anything else, the average is used. Default is "yes".}

\item{.criticalVal}{The critical hit multiplier. Default is 1.5 (2 in Generation V).}

\item{.PBVal}{The Parental Bond multiplier. Default is 0.25 (0.5 in Generation VI).}
}
\value{
The calculated damage.
}
\description{
This function calculates the damage inflicted during a Pokemon battle.
It uses the formula from Generation V and onward.
}
\note{
this function does not handle the specific conditions mentioned (like Reflect, Light Screen, Explosion, Selfdestruct, etc.).
You would need to handle these conditions before passing the values to this function
}
\examples{
# From https://bulbapedia.bulbagarden.net/wiki/Damage#Example
pokemon_damage_V(75, 123, 163, 65, stab = TRUE, effectiveness = 4, random = "minmax")
pokemon_damage_V(75, 123, 163, 71, stab = TRUE, effectiveness = 4, critical = TRUE, random = "minmax")
}
