% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\name{process_pokemon}
\alias{process_pokemon}
\title{Checks if a given xml_node type of tag exists}
\usage{
process_pokemon(path = "E:\\\\html\\\\", scrape = FALSE)
}
\arguments{
\item{x}{The object to search}

\item{tag}{The tag to extract form the set}

\item{lst}{The list to modify}

\item{level}{The current level of recursion}

\item{tags}{The tags to move down (Note: this is not built to move things up)}
}
\value{
TRUE if it exists FALSE otherwise

A list of
matched_elements that have the matching xml_nodes
other_elements that have the updated list without the matched elements

The modified list
}
\description{
Checks if a given xml_node type of tag exists
}
\examples{
new_node = function(tag, val){
xml2::read_html(glue::glue("<{tag}>{val}</{tag}>"))|>
   xml2::xml_child()|>
   xml2::xml_child()
}

# example data
t = list(
   list(
       list(
           new_node("h2", "hi"),
           new_node("div", "hi")
       )
	  )
)
# > [[1]]
# > [[1]][[1]]
# > [[1]][[1]][[1]]
# > {html_node}
# > <h2>
# >
# > 	[[1]][[1]][[2]]
# > {html_node}
# > <div>
v = t|>
	recursive_xml_map2(relevel_matched_node)
# > [[1]]
# > [[1]][[1]]
# > {html_node}
# > <h2>
# >
# > [[1]][[2]]
# > [[1]][[2]][[1]]
# > {html_node}
# > <div>
}
