% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random.R
\name{randomPokemonGen}
\alias{randomPokemonGen}
\title{Returns a random Pokemon Generator}
\usage{
randomPokemonGen(..., p = English, replace = FALSE)
}
\arguments{
\item{p}{What column to return (Ndex, English, form, form2, type, type2 can be used)}

\item{replace}{Should sampling be with replacement?}

\item{type}{Filter on type}

\item{type2}{Filter on type2}

\item{swap}{Allow type to match type2 and vicea-versa}

\item{data}{Override the data (Must have type and type2 (If type 2 is not ANY or swap is TRUE))}
}
\description{
Returns a random Pokemon Generator
}
