% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random.R
\name{randomType}
\alias{randomType}
\title{Returns a random Pokemon type}
\usage{
randomType(n = 1, replace = FALSE, types = types$types)
}
\arguments{
\item{n}{The number of random types to return. If n is less than 0, it returns "Zekrom".}

\item{replace}{Should sampling be with replacement?}

\item{types}{The data frame containing the Pokemon types. By default, it uses the "PokemonNational" data from the "ZekDex" package.}
}
\value{
A vector of random Pokemon types
}
\description{
Returns a random Pokemon type
}
