% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{recursive_xml_map}
\alias{recursive_xml_map}
\title{Recursively modifies a list}
\usage{
recursive_xml_map(lst, fun)
}
\arguments{
\item{lst}{A list to operate on}

\item{fun}{A function to operate on the end of the lst}
}
\value{
The modified lst
}
\description{
Recursively modifies a list
}
