% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{recursive_xml_map2}
\alias{recursive_xml_map2}
\title{Recursively modifies a list on each list}
\usage{
recursive_xml_map2(lst, fn, level = 1)
}
\arguments{
\item{lst}{A list to operate on}

\item{fn}{A function to operate on chunks of the lst

\if{html}{\out{<div class="sourceCode">}}\preformatted{    x: the current chunk of the list
    level: How deep recursion is currently
    return: A modified list chunk used in the next recursions and returned by recursive_xml_map2
}\if{html}{\out{</div>}}}

\item{level}{How deep recursion is currently, starts at 1 by default.  Increments every recursion}
}
\value{
The modified lst
}
\description{
Recursively modifies a list on each list
}
