% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{regionalForm}
\alias{regionalForm}
\title{Regional Form Function}
\usage{
regionalForm(re = FALSE, ext = FALSE, wrap = TRUE)
}
\arguments{
\item{re}{A logical value. If TRUE, the function returns a regex pattern.
If FALSE, it returns a list. Default is FALSE.}

\item{ext}{A logical value. If TRUE, the function includes extended regional
forms in the output. Default is FALSE.}
}
\value{
If re is TRUE, a character string representing a regex pattern of
Pokémon regional forms. If re is FALSE, a character vector of Pokémon
regional forms.
}
\description{
This function provides a list or a regex pattern of Pokémon regional forms.
}
\examples{
regionalForm(re = TRUE, ext = TRUE)
regionalForm(re = FALSE, ext = FALSE)

}
