% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\name{relevel_matched_node}
\alias{relevel_matched_node}
\title{Relevels the structure of a list of xml_nodes and xml_nodesets}
\usage{
relevel_matched_node(lst, level = 1)
}
\arguments{
\item{lst}{The list to modify}

\item{level}{The current level of recursion}

\item{tags}{The tags to move down (Note: this is not built to move things up)}
}
\value{
The modified list
}
\description{
Relevels the structure of a list of xml_nodes and xml_nodesets
}
\examples{
new_node = function(tag, val){
xml2::read_html(glue::glue("<{tag}>{val}</{tag}>"))|>
   xml2::xml_child()|>
   xml2::xml_child()
}

# example data
t = list(
   list(
       list(
           new_node("h2", "hi"),
           new_node("div", "hi")
       )
	  )
)
# > [[1]]
# > [[1]][[1]]
# > [[1]][[1]][[1]]
# > {html_node}
# > <h2>
# >
# > 	[[1]][[1]][[2]]
# > {html_node}
# > <div>
v = t|>
	recursive_xml_map2(relevel_matched_node)
# > [[1]]
# > [[1]][[1]]
# > {html_node}
# > <h2>
# >
# > [[1]][[2]]
# > [[1]][[2]][[1]]
# > {html_node}
# > <div>
}
