% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{removeA}
\alias{removeA}
\title{Replaces all accented characters in a string with their unaccented counterparts}
\usage{
removeA(str)
}
\arguments{
\item{str}{The string in which to replace accented characters}
}
\value{
A string with all accented characters replaced by their unaccented counterparts
}
\description{
Replaces all accented characters in a string with their unaccented counterparts
}
