% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{save_data}
\alias{save_data}
\title{Save Data to CSV and/or RDA}
\usage{
save_data(
  sym,
  root = stop("'root' must be specified"),
  file = stop("'file' must be specified"),
  csv = TRUE,
  rda = TRUE
)
}
\arguments{
\item{sym}{The name of the data frame to save.}

\item{root}{The root directory where the file should be saved.}

\item{file}{The name of the file (without extension).}

\item{csv}{Logical, should the data be saved as a CSV file? Default is TRUE.}

\item{rda}{Logical, should the data be saved as an RDA file? Default is TRUE.}
}
\description{
This function saves a given data frame to a CSV and/or RDA file.
}
