% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\name{scrape_pokemon}
\alias{scrape_pokemon}
\title{Scrape Pokémon Data from Bulbapedia}
\usage{
scrape_pokemon(path = "E:\\\\html\\\\")
}
\arguments{
\item{path}{The directory where the RDS files will be saved. Default is "E:\\html\\".}
}
\value{
None. This function does not return anything but saves the HTML content of each Pokémon's page as an RDS file in the specified directory.
}
\description{
This function scrapes Pokémon data from Bulbapedia and saves the HTML content of each Pokémon's page as an RDS file in the specified directory.
}
