% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{split_each}
\alias{split_each}
\title{Splits a list at each specified tag}
\usage{
split_each(lst, tags = c("h2", "h3", "h4"))
}
\arguments{
\item{lst}{A list to split.}

\item{tags}{A vector of tags at which to split the list.}
}
\value{
A list of lists, split at each specified tag.
}
\description{
This function splits a list at each specified tag. It returns a list of lists, split at each specified tag.
}
\examples{
split_each(list("h1", "h2", "h3", "h4"), c("h2", "h3", "h4"))
}
