% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{stab}
\alias{stab}
\title{Compute STAB Multiplier for an Attack}
\usage{
stab(effectiveness, stab_mult = stab_multiplyer$Standard)
}
\arguments{
\item{effectiveness}{Numeric value representing the effectiveness of an attack.}

\item{stab_mult}{Numeric value representing the STAB multiplier. Default is stab_multiplyer$Standard.}
}
\value{
Numeric value representing the effectiveness of the attack with the STAB multiplier applied.
}
\description{
This function computes the Same Type Attack Bonus (STAB) multiplier for an attack.
It is to be used with the \code{effectiveness} and \code{effectivenessPokemon} functions instead of passing a parameter.
}
\examples{
effectivenessPokemon("Fire", "Zekrom")|>
  stab()
}
