% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\docType{data}
\name{stab_multiplyer}
\alias{stab_multiplyer}
\title{List of STAB Multipliers}
\format{
list
}
\source{
Bulbapedia
}
\usage{
stab_multiplyer
}
\description{
This list contains various Same Type Attack Bonus (STAB) multipliers used in different Pokémon games.
Each multiplier represents a different game or game mechanic.
}
\examples{
stab_multiplyer$Standard
stab_multiplyer$Adaptability
stab_multiplyer$LegendsArceus
stab_multiplyer$Terastallized
stab_multiplyer$AdaptabilityTerastallized
stab_multiplyer$MysteryDungeon
stab_multiplyer$GO
stab_multiplyer$GO2017
}
\keyword{datasets}
