% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats}
\alias{stats}
\title{Pokemon Stats}
\format{
A data frame with 11 variables:
\describe{
\item{ndex}{\code{integer} National Pokédex number. This is the unique identifier for each Pokémon in the National Pokédex.}
\item{name}{\code{character} Pokémon name. This is the common name used to identify a particular Pokémon.}
\item{HP}{\code{integer} Hit Points. This determines how much damage a Pokémon can receive before fainting.}
\item{Attack}{\code{integer} Attack strength. This determines how much damage a Pokémon can deal with physical moves.}
\item{Defense}{\code{integer} Defensive strength. This determines how much damage a Pokémon can resist from physical moves.}
\item{Speed}{\code{integer} Speed. This determines the order of Pokémon that can act in battle.}
\item{Special}{\code{integer} Special attack or special defense strength. This determines how much damage a Pokémon can deal with special moves or resist from special moves. Note that this stat is combined into one in Generation I.}
\item{Total}{\code{integer} Total of all stats. This is the sum of all the base stats of a Pokémon.}
\item{Average}{\code{numeric} Average of all stats. This is the average value of all the base stats of a Pokémon.}
\item{Gen}{\code{factor} Generation. This indicates the generation of games in which the Pokémon was introduced.}
\item{SpAttack}{\code{integer} Special Attack strength. This determines how much damage a Pokémon can deal with special moves. This stat is not applicable for Generation I as 'Special' stat is used instead.}
\item{SpDefense}{\code{integer} Special Defense strength. This determines how much damage a Pokémon can resist from special moves. This stat is not applicable for Generation I as 'Special' stat is used instead.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_I)}{List of Pokémon by base stats (Generation I). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_II-V)}{List of Pokémon by base stats (Generation II-V). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_VI)}{List of Pokémon by base stats (Generation VI). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_VII)}{List of Pokémon by base stats (Generation VII). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_VIII)}{List of Pokémon by base stats (Generation VIII). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_IX)}{List of Pokémon by base stats (Generation IX). In Bulbapedia.}
}
\usage{
stats
}
\description{
This dataset contains the base stats of Pokémon, including various forms and sizes of Pokémon. Base stats are an important defining characteristic of each Pokémon species. They determine the potential of each Pokémon in terms of their HP (Hit Points), Attack, Defense, Speed, and Special abilities. These stats can be influenced by other factors such as level, nature, and effort values.
}
\examples{
data(stats)
}
\keyword{datasets}
