% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statsWide}
\alias{statsWide}
\title{Pokemon Stats (Wide Format)}
\format{
A data frame with 49 variables:
\describe{
\item{\code{ndex}}{\code{integer} National Pokédex number. This is the unique identifier for each Pokémon in the National Pokédex.}
\item{\code{name}}{\code{character} Pokémon name. This is the common name used to identify a particular Pokémon.}
\item{\code{Mega}}{\code{character} The Mega form. These are special forms that certain Pokémon can take.}
\item{\code{MegaOrPrimal}}{\code{factor} Indicates if the Pokémon is in a Mega or Primal form. These are special forms that certain Pokémon can take.}
\item{\code{regional}}{\code{factor} The regional form of the Pokémon. This refers to the variations of Pokémon that appear differently and have different attributes compared to their counterparts in other regions.}
\item{\code{form}}{\code{character} Form of the Pokemon. This refers to the different physical forms that Pokémon can take. NA if the Pokemon does not have a special form.}
\item{\code{HP_I}}{\code{integer} Hit Points for Generation I. This determines how much damage a Pokémon can receive before fainting.}
\item{\code{HP_VI}}{\code{integer} Hit Points for Generation VI. This determines how much damage a Pokémon can receive before fainting.}
\item{\code{HP_VII}}{\code{integer} Hit Points for Generation VII. This determines how much damage a Pokémon can receive before fainting.}
\item{\code{HP_VIII}}{\code{integer} Hit Points for Generation VIII. This determines how much damage a Pokémon can receive before fainting.}
\item{\code{HP_IX}}{\code{integer} Hit Points for Generation IX. This determines how much damage a Pokémon can receive before fainting.}
\item{\code{HP_II-V}}{\code{integer} Hit Points for Generations II to V. This determines how much damage a Pokémon can receive before fainting.}
\item{\code{Attack_I}}{\code{integer} Attack strength for Generation I. This determines how much damage a Pokémon can deal with physical moves.}
\item{\code{Attack_VI}}{\code{integer} Attack strength for Generation VI. This determines how much damage a Pokémon can deal with physical moves.}
\item{\code{Attack_VII}}{\code{integer} Attack strength for Generation VII. This determines how much damage a Pokémon can deal with physical moves.}
\item{\code{Attack_VIII}}{\code{integer} Attack strength for Generation VIII. This determines how much damage a Pokémon can deal with physical moves.}
\item{\code{Attack_IX}}{\code{integer} Attack strength for Generation IX. This determines how much damage a Pokémon can deal with physical moves.}
\item{\code{Attack_II-V}}{\code{integer} Attack strength for Generations II to V. This determines how much damage a Pokémon can deal with physical moves.}
\item{\code{Defense_I}}{\code{integer} Defensive strength for Generation I. This determines how much damage a Pokémon can resist from physical moves.}
\item{\code{Defense_VI}}{\code{integer} Defensive strength for Generation VI. This determines how much damage a Pokémon can resist from physical moves.}
\item{\code{Defense_VII}}{\code{integer} Defensive strength for Generation VII. This determines how much damage a Pokémon can resist from physical moves.}
\item{\code{Defense_VIII}}{\code{integer} Defensive strength for Generation VIII. This determines how much damage a Pokémon can resist from physical moves.}
\item{\code{Defense_IX}}{\code{integer} Defensive strength for Generation IX. This determines how much damage a Pokémon can resist from physical moves.}
\item{\code{Defense_II-V}}{\code{integer} Defensive strength for Generations II to V. This determines how much damage a Pokémon can resist from physical moves.}
\item{\code{Speed_I}}{\code{integer} Speed for Generation I. This determines the order of Pokémon that can act in battle.}
\item{\code{Speed_VI}}{\code{integer} Speed for Generation VI. This determines the order of Pokémon that can act in battle.}
\item{\code{Speed_VII}}{\code{integer} Speed for Generation VII. This determines the order of Pokémon that can act in battle.}
\item{\code{Speed_VIII}}{\code{integer} Speed for Generation VIII. This determines the order of Pokémon that can act in battle.}
\item{\code{Speed_IX}}{\code{integer} Speed for Generation IX. This determines the order of Pokémon that can act in battle.}
\item{\code{Speed_II-V}}{\code{integer} Speed for Generations II to V. This determines the order of Pokémon that can act in battle.}
\item{\code{Special_I}}{\code{integer} Special attack or special defense strength for Generation I. This determines how much damage a Pokémon can deal with special moves or resist from special moves. Note that this stat is combined into one in Generation I.}
\item{\code{Total_I}}{\code{integer} Total of all stats for Generation I. This is the sum of all the base stats of a Pokémon.}
\item{\code{Total_VI}}{\code{integer} Total of all stats for Generation VI. This is the sum of all the base stats of a Pokémon.}
\item{\code{Total_VII}}{\code{integer} Total of all stats for Generation VII. This is the sum of all the base stats of a Pokémon.}
\item{\code{Total_VIII}}{\code{integer} Total of all stats for Generation VIII. This is the sum of all the base stats of a Pokémon.}
\item{\code{Total_IX}}{\code{integer} Total of all stats for Generation IX. This is the sum of all the base stats of a Pokémon.}
\item{\code{Total_II-V}}{\code{integer} Total of all stats for Generations II to V. This is the sum of all the base stats of a Pokémon.}
\item{\code{Average_I}}{\code{numeric} Average of all stats for Generation I. This is the average value of all the base stats of a Pokémon.}
\item{\code{Average_VI}}{\code{numeric} Average of all stats for Generation VI. This is the average value of all the base stats of a Pokémon.}
\item{\code{Average_VII}}{\code{numeric} Average of all stats for Generation VII. This is the average value of all the base stats of a Pokémon.}
\item{\code{Average_VIII}}{\code{numeric} Average of all stats for Generation VIII. This is the average value of all the base stats of a Pokémon.}
\item{\code{Average_IX}}{\code{numeric} Average of all stats for Generation IX. This is the average value of all the base stats of a Pokémon.}
\item{\code{Average_II-V}}{\code{numeric} Average of all stats for Generations II to V. This is the average value of all the base stats of a Pokémon.}
\item{\code{SpAttack_VI}}{\code{integer} Special Attack strength for Generation VI. This determines how much damage a Pokémon can deal with special moves.}
\item{\code{SpAttack_VII}}{\code{integer} Special Attack strength for Generation VII. This determines how much damage a Pokémon can deal with special moves.}
\item{\code{SpAttack_VIII}}{\code{integer} Special Attack strength for Generation VIII. This determines how much damage a Pokémon can deal with special moves.}
\item{\code{SpAttack_IX}}{\code{integer} Special Attack strength for Generation IX. This determines how much damage a Pokémon can deal with special moves.}
\item{\code{SpAttack_II-V}}{\code{integer} Special Attack strength for Generations II to V. This determines how much damage a Pokémon can deal with special moves.}
\item{\code{SpDefense_VI}}{\code{integer} Special Defense strength for Generation VI. This determines how much damage a Pokémon can resist from special moves.}
\item{\code{SpDefense_VII}}{\code{integer} Special Defense strength for Generation VII. This determines how much damage a Pokémon can resist from special moves.}
\item{\code{SpDefense_VIII}}{\code{integer} Special Defense strength for Generation VIII. This determines how much damage a Pokémon can resist from special moves.}
\item{\code{SpDefense_IX}}{\code{integer} Special Defense strength for Generation IX. This determines how much damage a Pokémon can resist from special moves.}
\item{\code{SpDefense_II-V}}{\code{integer} Special Defense strength for Generations II to V. This determines how much damage a Pokémon can resist from special moves.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_I)}{List of Pokémon by base stats (Generation I). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_II-V)}{List of Pokémon by base stats (Generation II-V). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_VI)}{List of Pokémon by base stats (Generation VI). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_VII)}{List of Pokémon by base stats (Generation VII). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_VIII)}{List of Pokémon by base stats (Generation VIII). In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_base_stats_(Generation_IX)}{List of Pokémon by base stats (Generation IX). In Bulbapedia.}
}
\usage{
statsWide
}
\description{
It contains the base stats of Pokémon, including various forms and sizes of Pokémon, across different generations.
Base stats are an important defining characteristic of each Pokémon species.
They determine the potential of each Pokémon in terms of their HP (Hit Points),
Attack, Defense, Speed, and Special abilities.
These stats can be influenced by other factors such as level, nature, and effort values.
}
\details{
This dataset is a wide format version of the \code{stats} dataset.
}
\examples{
data(statsWide)
}
\keyword{datasets}
