% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{str_before_match}
\alias{str_before_match}
\title{Returns the string before a match}
\usage{
str_before_match(str, test, ..., .open = "(", .close = ")?")
}
\arguments{
\item{str}{The string to resize.}

\item{test}{The string to test.}

\item{...}{Additional arguments to be passed to the \code{regex} function.}

\item{.open}{The opening character(s) for wrapping in the .recursive_wrap function. Default is "(".}

\item{.close}{The closing character(s) for wrapping in the .recursive_wrap function. Default is ")?".}
}
\value{
The string before the match.
}
\description{
This function returns the string before a match. It uses the \code{str_longest_match} function to find the match and then returns the string before it.
}
\examples{
ZekDex:::str_before_match("Zekrom", "Zek")
}
