% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{str_longest_match}
\alias{str_longest_match}
\title{Returns the shortest string before a match}
\usage{
str_longest_match(str, test, ..., .open = "(", .close = ")?")
}
\arguments{
\item{str}{The string to resize.}

\item{test}{The string to test.}
}
\value{
The shortest string not matching test.
}
\description{
This function returns the shortest string not matching a test string. It uses a recursive wrap function to match the strings.
}
\examples{
ZekDex:::str_longest_match("Zekrom", "Zek")
}
