% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\docType{data}
\name{type_colors}
\alias{type_colors}
\title{Generate a list of color palettes for Pokémon types}
\format{
An object of class \code{list} of length 2.
}
\usage{
type_colors
}
\value{
A list of named vectors representing color palettes for Pokémon types.
}
\description{
This function generates a list of color palettes for Pokémon types. The list contains two named vectors: 'new' and 'old', each representing a different color palette.
}
\keyword{datasets}
