% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{types}
\alias{types}
\title{Pokemon Types Dataset}
\format{
A data frame with one column:
\describe{
\item{\code{types}}{\code{factor} The types of Pokemon. Possible types include \code{Grass}, \code{Fire}, \code{Water}, \code{Bug}, \code{Normal}, \code{Dark}, \code{Poison}, \code{Electric}, \code{Ground}, \code{Ice}, \code{Fairy}, \code{Steel}, \code{Fighting}, \code{Psychic}, \code{Rock}, \code{Ghost}, \code{Dragon}, \code{Flying}.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/Type}{Bulbapedia Type}
}
\usage{
types
}
\description{
This dataset contains information about various Pokemon types.
}
\examples{
data(types)
}
\keyword{datasets}
