% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight}
\alias{weight}
\title{Pokemon Weight Dataset}
\format{
A data frame with 8 variables:
\describe{
\item{name}{\code{character} Name of the Pokemon}
\item{form}{\code{character} The form of the Pokemon, NA if it does not have another form}
\item{lbs}{\code{numeric} Weight of the Pokemon in pounds}
\item{kg}{\code{numeric} Weight of the Pokemon in kilograms}
\item{lightest}{\code{integer} Rank of the Pokemon from lightest (1) to heaviest (1181)}
\item{heaviest}{\code{integer} Rank of the Pokemon from heaviest (1) to lightest (1181)}
\item{regional}{\code{character} The regional form of the Pokemon, NA if it does not have a regional form}
\item{MegaOrPrimal}{\code{character} Is the Pokémon in a Mega or Primal form}
}
}
\source{
Bulbapedia
}
\usage{
weight
}
\description{
This dataset contains information about the weight of various Pokemon in both pounds (lbs) and kilograms (kg).
It also includes details about the lightest and heaviest Pokemon, their regional forms, and whether they have a Mega or Primal form.
}
\keyword{datasets}
