% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catchRate}
\alias{catchRate}
\title{Catch Rate Dataset}
\format{
A data frame with 5 variables:
\describe{
\item{\code{ndex}}{\code{integer} The National Pokédex number of the Pokémon. This is the unique identifier for each Pokémon in the National Pokédex.}
\item{\code{name}}{\code{character} The name of the Pokémon. This is the common name used to identify a particular Pokémon.}
\item{\code{catchRate}}{\code{integer} The catch rate of the Pokémon. This is a value that represents the probability of capturing the Pokémon. Higher values indicate a higher probability of capture.}
\item{\code{note}}{\code{character} Additional notes about the catch rate of the Pokémon. This may include information about specific conditions or factors that can influence the catch rate.}
\item{\code{hdex}}{\code{integer} The Pokémon Legends: Arceus Pokédex number. This is the unique identifier for each Pokémon in the Pokémon Legends: Arceus Pokédex.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_catch_rate}{Bulbapedia Catch Rate}
}
\usage{
catchRate
}
\description{
The catch rate is a statistic that determines the probability of capturing a Pokémon.
A higher catch rate means that the Pokémon is easier to catch.

In the context of Pokémon Legends: Arceus, many Pokémon were given heightened catch rates to facilitate catching them
outside of battle.
}
\details{
This dataset provides information about the catch rate of each Pokémon.
}
\examples{
data(catchRate)
}
\keyword{datasets}
