% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{.recursive_wrap}
\alias{.recursive_wrap}
\title{Recursively wraps a string with specified opening and closing characters}
\usage{
.recursive_wrap(str, .open = "(", .close = ")?")
}
\arguments{
\item{str}{The string to be wrapped.}

\item{.open}{The opening character(s) for wrapping. Default is "(".}

\item{.close}{The closing character(s) for wrapping. Default is ")?".}
}
\value{
A string that is recursively wrapped with the specified opening and closing characters.
}
\description{
This function recursively wraps a string with specified opening and closing characters.
It returns a string that is recursively wrapped with the specified opening and closing characters.
}
\examples{
ZekDex:::.recursive_wrap("Zekrom")
}
