% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{family}
\alias{family}
\title{Pokemon Evolution Dataset}
\format{
A data frame with the following variables:
\describe{
\item{family}{Character. The family to which the Pokemon belongs.}
\item{base}{Character. The prior form of the pokemon or the base form}
\item{form}{Character. The specific form of the Pokemon. NA if not applicable.}
\item{method}{Character. The method of evolution.}
\item{evo}{Character. The next stage of the Pokemon.}
\item{evoForm}{Character. The form of next stage of the Pokemon. NA if not applicable.}
\item{transition}{Character. The transition stage of the Pokemon.}
\item{first}{Character. The first form of the Pokemon.}
\item{firstForm}{Character. The specific first form of the Pokemon. NA if not applicable.}
\item{regional}{Character. The regional form of the Pokemon. NA if not applicable.}
\item{evoRegional}{Character. The regional evolved form of the Pokemon. NA if not applicable.}
\item{firstRegional}{Character. The regional first form of the Pokemon. NA if not applicable.}
}
}
\source{
Bulbapedia
}
\usage{
family
}
\description{
This dataset contains evolution and family information.
}
\keyword{datasets}
