% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{familyLong}
\alias{familyLong}
\title{Pokemon Dataset}
\format{
A data frame with the following variables:
\describe{
\item{family.x}{\code{character} The family of the Pokemon}
\item{transition}{\code{character} The transition stage of the Pokemon}
\item{first}{\code{character} The first form of the Pokemon}
\item{firstForm}{\code{character} The first form of the Pokemon if it has multiple forms}
\item{firstRegional}{\code{character} The first regional form of the Pokemon}
\item{name}{\code{character} The name of the Pokemon}
\item{form}{\code{character} The form of the Pokemon}
\item{regional}{\code{character} The regional form of the Pokemon}
\item{stage}{\code{character} The stage of the Pokemon}
\item{family.y}{\code{character} The legendary family of the Pokemon}
\item{size}{\code{character} The size of the legendary family}
}
}
\source{
Bulbapedia
}
\usage{
familyLong
}
\description{
This dataset contains evolution and family information.
}
\keyword{datasets}
