% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evos.R
\name{gen_evos}
\alias{gen_evos}
\title{Generate Pokémon Evolution Data}
\usage{
gen_evos(
  write = FALSE,
  path = system.file("data/PokemonEvolution.csv", package = "ZekDex")
)
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the data frame to a csv file. Default is \code{FALSE}.}

\item{path}{The path where the csv file will be written if \code{write = TRUE}. Default is \code{system.file("data/PokemonEvolution.csv", package = "ZekDex")}.}
}
\value{
A data frame of Pokémon evolution data.
}
\description{
This function scrapes Pokémon evolution data from Bulbapedia and returns a data frame. If \code{write = TRUE}, it also writes the data frame to a csv file.
}
\examples{
gen_evos()
}
