% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lang.R
\name{gen_lang}
\alias{gen_lang}
\title{Generate the National Dex with Other Language Data}
\usage{
gen_lang(
  write = FALSE,
  path = system.file("data/PokemonLang.csv", package = "ZekDex")
)
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the tibble to a csv file. Default is \code{FALSE}.}

\item{path}{The path where the csv file will be written if \code{write = TRUE}. Default is \code{system.file("data/PokemonLang.csv", package = "ZekDex")}.}
}
\value{
A tibble of the National Dex with language information.
}
\description{
This function generates a tibble of the National Dex with language information by reading the 'PokemonNational.csv' file and scraping data from Bulbapedia. It extracts language links and reads the HTML pages to compute the languages. It cleans the tables from each language and combines them into one dataframe. If \code{write = TRUE}, it also writes the tibble to a csv file.
}
