% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Natational.R
\name{gen_national}
\alias{gen_national}
\title{Generate the National Dex}
\usage{
gen_national(
  write = FALSE,
  path = system.file("data/PokemonNational.csv", package = "ZekDex")
)
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the tibble to a csv file. Default is \code{FALSE}.}

\item{path}{The path where the csv file will be written if \code{write = TRUE}. Default is \code{system.file("data/PokemonNational.csv", package = "ZekDex")}.}
}
\value{
A tibble of the National Dex.
}
\description{
This function generates a tibble of the National Dex by reading the 'PokemonNational.csv' file and scraping data from Bulbapedia. It extracts the tables from the HTML content, cleans the column names of the dataframe, and combines all dataframes into one. If \code{write = TRUE}, it also writes the tibble to a csv file.
}
\examples{
gen_national()
}
