% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regional.R
\name{gen_reginal}
\alias{gen_reginal}
\title{Generate Regional Pokedex}
\usage{
gen_reginal(write = FALSE)
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the tibble to a csv file. Default is \code{FALSE}.}
}
\value{
A tibble of regional dexes.
}
\description{
This function generates a tibble of the regional Pokedex by scraping data from Bulbapedia. It extracts the URLs of the regional Pokedex pages from the HTML, reads the HTML of each regional Pokedex page, and combines the regional Pokedex data with the national Pokedex data.
}
