% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{gen_stats}
\alias{gen_stats}
\title{Generate Pokémon Statistics}
\usage{
gen_stats(
  write = FALSE,
  path = system.file("data/PokemonStats.csv", package = "ZekDex"),
  pathWide = system.file("data/PokemonStatsWide.csv", package = "ZekDex")
)
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the data frames to csv files. Default is \code{FALSE}.}
}
\value{
A list of two data frames: PokemonStats (long format) and PokemonStatsWide (wide format).
}
\description{
This function scrapes Pokémon statistics from Bulbapedia and returns a list of two data frames: one in long format and one in wide format. If \code{write = TRUE}, it also writes these data frames to csv files.
}
\examples{
gen_stats()
}
