% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{gen_type}
\alias{gen_type}
\title{Generates Pokémon Types from Scratch}
\usage{
gen_type(
  write = FALSE,
  path = system.file("data/PokemonTypes.csv", package = "ZekDex")
)
}
\arguments{
\item{write}{Logical, if \code{TRUE}, writes the tibble to a csv file. Default is \code{FALSE}.}

\item{path}{The path where the csv file will be written if \code{write = TRUE}. Default is \code{system.file("data/PokemonTypes.csv", package = "ZekDex")}.}
}
\value{
A tibble with a single column named 'types' containing unique Pokémon types.
}
\description{
This function generates a tibble of Pokémon types by reading the 'PokemonNational.csv' file and extracting the 'type' and 'type2' columns. It removes duplicate and NA values to create a unique list of Pokémon types. If \code{write = TRUE}, it also writes the tibble to a csv file.
}
