% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups}
\alias{groups}
\title{Legendary and Mythical Pokémon Groupings}
\format{
A data frame with 3 variables:
\describe{
\item{\code{family}}{\code{character} The family grouping of the Pokémon. This refers to the thematic or characteristic-based group that the Pokémon belongs to. Examples include 'Tao trio', 'Super-ancient Pokémon', 'Paradox duo', etc.}
\item{\code{name}}{\code{character} The name of the individual Pokémon within the family grouping. This is the common name used to identify a particular Pokémon.}
\item{\code{size}}{\code{character} The size category of the family grouping. This refers to the number of Pokémon in the family grouping. Categories include 'duos', 'trios', 'quartets', and 'quintets'.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/Groups_of_Legendary_and_Mythical_Pok\%C3\%A9mon}{Groups of Legendary and Mythical Pokémon. In Bulbapedia.}
}
\usage{
groups
}
\description{
This dataset contains information about Legendary and Mythical Pokémon,
which are a special group of Pokémon that are incredibly rare and often very powerful.
These Pokémon are categorized by their family groupings and sizes.
Family groupings are often based on common themes or characteristics among the Pokémon.
}
\details{
For example, the 'Tao trio' refers to Reshiram, Zekrom, and Kyurem,
which are associated with the concept of yin, yang, and wuji.

The 'Super-ancient Pokémon' refers to Kyogre, Groudon, and Rayquaza,
which are ancient Pokémon with powerful weather abilities.

The 'Paradox duo' refers to Koraidon and Miraidon, which are associated with different timelines.
}
\examples{
data(groups)
}
\keyword{datasets}
