% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{languages}
\alias{languages}
\title{Pokemon National Dex Information}
\format{
A data frame with 37 variables:
\describe{
\item{ndex}{National Pokedex number. Integer.}
\item{name}{Name of the Pokemon. Character.}
\item{form}{Form of the Pokemon. Character. NA if the Pokemon does not have a special form.}
\item{regional}{Regional form of the Pokemon. Character. NA if the Pokemon does not have a regional form.}
\item{type}{Primary type of the Pokemon. Character.}
\item{type2}{Secondary type of the Pokemon. Character. NA if the Pokemon does not have a secondary type.}
\item{generation}{The generation the Pokemon was introduced in. Integer.}
\item{isLegendary}{Whether the Pokemon is Legendary. Logical.}
\item{isMythical}{Whether the Pokemon is Mythical. Logical.}
\item{isUltraBeast}{Whether the Pokemon is an Ultra Beast. Logical.}
\item{family}{Legendary / Mythical family group of the Pokemon. Character. NA if the Pokemon does not belong to a family group.}
\item{size}{Legendary / Mythical family size. NA if the Pokemon does not belong to a family group.}
\item{Japanese_Kana}{Japanese Kana name of the Pokemon. Character.}
\item{Japanese_Hepburn}{Japanese Hepburn name of the Pokemon. Character.}
\item{Japanese_Trademarked}{Japanese Trademarked name of the Pokemon. Character.}
\item{German_german}{German name of the Pokemon. Character.}
\item{French_french}{French name of the Pokemon. Character.}
\item{Spanish_spanish}{Spanish name of the Pokemon. Character.}
\item{Italian_italian}{Italian name of the Pokemon. Character.}
\item{Korean_Hangul}{Korean Hangul name of the Pokemon. Character.}
\item{Korean_Revised}{Korean Revised name of the Pokemon. Character.}
\item{Korean_McCune-Reischauer}{Korean McCune-Reischauer name of the Pokemon. Character.}
\item{Korean_Official}{Korean Official name of the Pokemon. Character.}
\item{Chinese_Traditional}{Chinese Traditional name of the Pokemon. Character.}
\item{Chinese_Simplified}{Chinese Simplified name of the Pokemon. Character.}
\item{Chinese_Pinyin}{Chinese Pinyin name of the Pokemon. Character.}
\item{Chinese_Yale}{Chinese Yale name of the Pokemon. Character.}
\item{Chinese_Former names}{Chinese Former names of the Pokemon. Character.}
\item{Brazilian_Portuguese_brazilian_portuguese}{Brazilian Portuguese name of the Pokemon. Character.}
\item{Turkish_turkish}{Turkish name of the Pokemon. Character.}
\item{Russian_Cyrillic}{Russian Cyrillic name of the Pokemon. Character.}
\item{Russian_Romanization}{Russian Romanization name of the Pokemon. Character.}
\item{Thai_Thai}{Thai name of the Pokemon. Character.}
\item{Thai_RTGS}{Thai RTGS name of the Pokemon. Character.}
\item{Hindi_Devan훮gar카}{Hindi Devan훮gar카 name of the Pokemon. Character.}
\item{Hindi_Transliteration}{Hindi Transliteration name of the Pokemon. Character.}
\item{Hindi_Romanization}{Hindi Romanization name of the Pokemon. Character.}
}
}
\source{
Bulbapedia
}
\usage{
languages
}
\description{
This dataset contains information about Pokemon from the National Dex.
}
\examples{
data(languages)
}
\keyword{datasets}
