% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{languages}
\alias{languages}
\title{Pokémon National Dex information in multiple languages}
\format{
A data frame with 37 variables:
\describe{
\item{\code{ndex}}{\code{integer} The National Pokédex number of the Pokémon. This is the unique identifier for each Pokémon in the National Pokédex.}
\item{\code{name}}{\code{character} The English name of the Pokémon.}
\item{\code{form}}{\code{character} The form of the Pokémon. This refers to the different physical forms that Pokémon can take. 'NA' if the Pokémon does not have a special form.}
\item{\code{regional}}{\code{factor} The regional form of the Pokémon. This refers to the variations of Pokémon that appear differently and have different attributes compared to their counterparts in other regions. 'NA' if the Pokémon does not have a regional form.}
\item{\code{type}}{\code{factor} The primary type of the Pokémon. Each Pokémon has a primary type which determines the Pokémon's strengths and weaknesses.}
\item{\code{type2}}{\code{factor} The secondary type of the Pokémon. Some Pokémon may have a secondary type, which further influences the Pokémon's strengths and weaknesses. 'NA' if the Pokémon does not have a secondary type.}
\item{\code{generation}}{\code{integer} The generation of the Pokémon. This refers to the generation of games in which the Pokémon was first introduced.}
\item{\code{isLegendary}}{\code{logical} A boolean indicating if the Pokémon is Legendary. Legendary Pokémon are a group of incredibly rare and often very powerful Pokémon.}
\item{\code{isMythical}}{\code{logical} A boolean indicating if the Pokémon is Mythical. Mythical Pokémon are a related but separate group of Pokémon, which are usually event-exclusive.}
\item{\code{isUltraBeast}}{\code{logical} A boolean indicating if the Pokémon is an Ultra Beast. Ultra Beasts is a term introduced in Generation VII to classify seven known Pokémon.}
\item{\code{family}}{\code{factor} Legendary / Mythical family group of the Pokemon. This refers to the group or lineage the Legendary or Mythical Pokémon belongs to. NA if the Pokemon does not belong to a family group.}
\item{\code{size}}{\code{factor} Legendary / Mythical family size. This refers to the size of the Legendary or Mythical family group the Pokémon belongs to. NA if the Pokemon does not belong to a family group.}
\item{\code{Japanese_Kana}}{\code{character} The Japanese Kana name of the Pokémon.}
\item{\code{Japanese_Hepburn}}{\code{character} The Japanese Hepburn name of the Pokémon.}
\item{\code{Japanese_Trademarked}}{\code{character} The Japanese Trademarked name of the Pokémon.}
\item{\code{German}}{\code{character} The German name of the Pokémon.}
\item{\code{French}}{\code{character} The French name of the Pokémon.}
\item{\code{Spanish}}{\code{character} The Spanish name of the Pokémon.}
\item{\code{Italian}}{\code{character} The Italian name of the Pokémon.}
\item{\code{Korean_Hangul}}{\code{character} The Korean Hangul name of the Pokémon.}
\item{\code{Korean_Revised}}{\code{character} The Korean Revised name of the Pokémon.}
\item{\code{Korean_McCune-Reischauer}}{\code{character} The Korean McCune-Reischauer name of the Pokémon.}
\item{\code{Korean_Official}}{\code{character} The Korean Official name of the Pokémon.}
\item{\code{Chinese_Traditional}}{\code{character} The Chinese Traditional name of the Pokémon.}
\item{\code{Chinese_Simplified}}{\code{character} The Chinese Simplified name of the Pokémon.}
\item{\code{Chinese_Pinyin}}{\code{character} The Chinese Pinyin name of the Pokémon.}
\item{\code{Chinese_Yale}}{\code{character} The Chinese Yale name of the Pokémon.}
\item{\code{Chinese_Former_names}}{\code{character} The Chinese Former names of the Pokémon.}
\item{\code{Brazilian_Portuguese}}{\code{character} The Brazilian Portuguese name of the Pokémon.}
\item{\code{Turkish}}{\code{character} The Turkish name of the Pokémon.}
\item{\code{Russian_Cyrillic}}{\code{character} The Russian Cyrillic name of the Pokémon.}
\item{\code{Russian_Romanization}}{\code{character} The Russian Romanization name of the Pokémon.}
\item{\code{Thai}}{\code{character} The Thai name of the Pokémon.}
\item{\code{Thai_RTGS}}{\code{character} The Thai RTGS name of the Pokémon.}
\item{\code{Hindi_Devanagari}}{\code{character} The Hindi Devanāgarī name of the Pokémon.}
\item{\code{Hindi_Transliteration}}{\code{character} The Hindi Transliteration name of the Pokémon.}
\item{\code{Hindi_Romanization}}{\code{character} The Hindi Romanization name of the Pokémon.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_National_Pok\%C3\%A9dex_number}{List of Pokémon by National Pokédex number. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Japanese_Pok\%C3\%A9mon_names}{List of Japanese Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_German_Pok\%C3\%A9mon_names}{List of German Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_French_Pok\%C3\%A9mon_names}{List of French Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Spanish_Pok\%C3\%A9mon_names}{List of Spanish Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Italian_Pok\%C3\%A9mon_names}{List of Italian Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Korean_Pok\%C3\%A9mon_names}{List of Korean Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Chinese_Pok\%C3\%A9mon_names}{List of Chinese Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Brazilian_Portuguese_Pok\%C3\%A9mon_names}{List of Brazilian Portuguese Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Turkish_Pok\%C3\%A9mon_names}{List of Turkish Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Russian_Pok\%C3\%A9mon_names}{List of Russian Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Thai_Pok\%C3\%A9mon_names}{List of Thai Pokémon names. In Bulbapedia.}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Hindi_Pok\%C3\%A9mon_names}{List of Hindi Pokémon names. In Bulbapedia.}
}
\usage{
languages
}
\description{
This dataset contains the names of Pokémon in various languages.
Each Pokémon has a unique name in each language, and these names often reflect the culture and
language patterns of their respective regions. The dataset includes names in multiple languages,
including but not limited to English, Japanese, German, French, Spanish, Italian, Korean, Chinese,
Brazilian Portuguese, Turkish, Russian, Thai, and Hindi. The language columns are in the format of
\verb{<Language>_<language subset>}. If there is no subset, it's \verb{<Language>}.
}
\examples{
data(languages)
}
\keyword{datasets}
