% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nationalDex}
\alias{nationalDex}
\title{National PokeDex}
\format{
A data frame with 12 variables:
\describe{
\item{ndex}{\code{integer} National Pokedex number.}
\item{name}{\code{character} Name of the Pokemon.}
\item{form}{\code{character} Form of the Pokemon. NA if the Pokemon does not have a special form.}
\item{regional}{\code{character} Regional form of the Pokemon. NA if the Pokemon does not have a regional form.}
\item{type}{\code{character} Primary type of the Pokemon.}
\item{type2}{\code{character} Secondary type of the Pokemon. NA if the Pokemon does not have a secondary type.}
\item{generation}{\code{integer} The generation the Pokemon was introduced in.}
\item{isLegendary}{\code{logical} Whether the Pokemon is Legendary.}
\item{isMythical}{\code{logical} Whether the Pokemon is Mythical.}
\item{isUltraBeast}{\code{logical} Whether the Pokemon is an Ultra Beast.}
\item{family}{\code{character} Legendary / Mythical family group of the Pokemon. NA if the Pokemon does not belong to a family group.}
\item{size}{\code{character} Legendary / Mythical family size. NA if the Pokemon does not belong to a family group.}
}
}
\source{
Bulbapedia
}
\usage{
nationalDex
}
\description{
Pokemon National dex information
}
\examples{
data(nationalDex)
}
\keyword{datasets}
