% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physicalAttr}
\alias{physicalAttr}
\title{Physical Attributes of Pokémon}
\format{
A data frame with 22 variables:
\describe{
\item{\code{ndex}}{\code{integer} The National Pokédex number of the Pokémon. This is the official numbering scheme that is used in every Pokémon game and media.}
\item{\code{name}}{\code{character} The name of the Pokémon. This is the common name used to identify a particular Pokémon.}
\item{\code{regional}}{\code{character} The regional variant of the Pokémon, if any. This refers to the variations of Pokémon that appear differently and have different attributes compared to their counterparts in other regions.}
\item{\code{type}}{\code{character} The primary type of the Pokémon. Each Pokémon has a primary type which determines the Pokémon's strengths and weaknesses.}
\item{\code{type2}}{\code{character} The secondary type of the Pokémon, if any. Some Pokémon may have a secondary type, which further influences the Pokémon's strengths and weaknesses.}
\item{\code{generation}}{\code{integer} The generation of the Pokémon. This refers to the generation of games in which the Pokémon was first introduced.}
\item{\code{isLegendary}}{\code{logical} A boolean indicating if the Pokémon is Legendary. Legendary Pokémon are a group of incredibly rare and often very powerful Pokémon.}
\item{\code{isMythical}}{\code{logical} A boolean indicating if the Pokémon is Mythical. Mythical Pokémon are a related but separate group of Pokémon, which are usually event-exclusive.}
\item{\code{isUltraBeast}}{\code{logical} A boolean indicating if the Pokémon is an Ultra Beast. Ultra Beasts is a term introduced in Generation VII to classify seven known Pokémon.}
\item{\code{family}}{\code{character} The family of the Pokémon. This refers to the group or lineage the Pokémon belongs to, often based on common themes or characteristics among the Pokémon.}
\item{\code{size}}{\code{character} The size of the Pokémon. This refers to the height and weight of the Pokémon.}
\item{\code{form}}{\code{character} The form of the Pokémon. This refers to the different physical forms that Pokémon can take.}
\item{\code{lbs}}{\code{numeric} The weight of the Pokémon in pounds.}
\item{\code{kg}}{\code{numeric} The weight of the Pokémon in kilograms.}
\item{\code{lightest}}{\code{integer} The rank of the Pokémon from lightest (1) to heaviest (1181).}
\item{\code{heaveiest}}{\code{integer} The rank of the Pokémon from heaviest (1) to lightest (1181).}
\item{\code{MegaOrPrimal}}{\code{character} Indicates if the Pokémon is in a Mega or Primal form. These are special forms that certain Pokémon can take.}
\item{\code{ft}}{\code{character} The height of the Pokémon in feet.}
\item{\code{m}}{\code{numeric} The height of the Pokémon in meters.}
\item{\code{inch}}{\code{integer} The height of the Pokémon in inches.}
\item{\code{smallest}}{\code{integer} The ranking of the Pokémon from smallest to biggest.}
\item{\code{bigest}}{\code{integer} The ranking of the Pokémon from biggest to smallest.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_weight}{Bulbapedia Weight}
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_height}{Bulbapedia Height}
}
\usage{
physicalAttr
}
\description{
This dataset provides a comprehensive overview of the physical attributes of Pokémon, including their type,
generation, weight, height, and other characteristics.
It includes information about each Pokémon's National Pokédex number, name, regional variant,
primary and secondary types, generation, legendary status, mythical status, ultra beast status,
family, size, form, weight in pounds and kilograms, lightest and heaviest weight rankings in the Pokédex,
Mega or Primal form status, height in feet, meters, and inches, and smallest and
biggest height rankings in the Pokédex.
}
\details{
This dataset can be useful for a variety of analyses, such as studying the distribution of Pokémon
weights and heights, identifying trends and patterns, or informing game strategies.
}
\examples{
data(physicalAttr)
}
\keyword{datasets}
