% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{pokemon_damage_II}
\alias{pokemon_damage_II}
\title{Pokemon Damage Calculation Generation II}
\usage{
pokemon_damage_II(
  level,
  attackStat,
  defenseStat,
  power,
  effectiveness,
  item = FALSE,
  critical = FALSE,
  TK = 1,
  weather = 1,
  badge = FALSE,
  stab = FALSE,
  moveMod = 1,
  doubleDmg = FALSE,
  random = "yes"
)
}
\arguments{
\item{level}{The level of the attacking Pokemon.}

\item{attackStat}{The effective Attack stat of the attacking Pokemon.}

\item{defenseStat}{The effective Defense stat of the target Pokemon.}

\item{power}{The power of the used move.}

\item{effectiveness}{The product of Type, and Type2.}

\item{item}{Logical. If TRUE, the attacker is holding an type-enhancing held item corresponding to the attack type. Default is FALSE.}

\item{critical}{Logical. If TRUE, it's a critical hit. Default is FALSE.}

\item{stab}{Logical. If TRUE, add the same type attack bonus.}

\item{random}{Character string. If "yes", a random uniformly distributed integer between 217 and 255 is used. If "minmax", the min and max values are returned. If anything else, the average is used. Default is "yes".}
}
\value{
The calculated damage.
}
\description{
This function calculates the damage inflicted during a Pokemon battle.
It uses the formula from Generation II.
}
\note{
this function does not handle the specific conditions mentioned (like Reflect, Light Screen, Explosion, Selfdestruct, etc.).
You would need to handle these conditions before passing the values to this function
}
