% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damage.R
\name{pokemon_damage_IV}
\alias{pokemon_damage_IV}
\title{Pokemon Damage Calculation Generation IV}
\usage{
pokemon_damage_IV(
  level,
  attackStat,
  defenseStat,
  power,
  effectiveness,
  burn = FALSE,
  screen = 1,
  targets = 1,
  weather = 1,
  FF = FALSE,
  critical = FALSE,
  item = FALSE,
  first = FALSE,
  stab = FALSE,
  SRF = FALSE,
  EB = FALSE,
  TL = FALSE,
  Berry = FALSE,
  random = "yes"
)
}
\arguments{
\item{level}{The level of the attacking Pokemon.}

\item{attackStat}{The effective Attack stat of the attacking Pokemon.}

\item{defenseStat}{The effective Defense stat of the target Pokemon.}

\item{power}{The power of the used move.}

\item{effectiveness}{The product of Type, and Type2.}

\item{burn}{Logical. If TRUE, the attacker is burned, its Ability is not Guts, and the used move is a physical move. Default is FALSE.}

\item{screen}{The screen multiplier. Default is 1.}

\item{targets}{The targets multiplier. Default is 1.}

\item{weather}{The weather multiplier. Default is 1.}

\item{FF}{Logical. If TRUE, the used move is Fire-type, and the attacker's Ability is Flash Fire that has been activated by a Fire-type move. Default is FALSE.}

\item{critical}{Logical. If TRUE, it's a critical hit. Default is FALSE.}

\item{item}{Logical. If TRUE, the attacker is holding an type-enhancing held item corresponding to the attack type. Default is FALSE.}

\item{first}{Logical. If TRUE, the used move was stolen with Me First. Default is FALSE.}

\item{stab}{Logical. If TRUE, add the same type attack bonus.}

\item{SRF}{Logical. If TRUE, the used move is super effective, the target's Ability is Solid Rock or Filter, and the attacker's Ability is not Mold Breaker. Default is FALSE.}

\item{EB}{Logical. If TRUE, the used move is super effective and the attacker is holding an Expert Belt. Default is FALSE.}

\item{TL}{Logical. If TRUE, the used move is not very effective and the attacker's Ability is Tinted Lens. Default is FALSE.}

\item{Berry}{Logical. If TRUE, the used move is super effective and the target is holding the Berry that weakens it, or Normal-type and the target is holding a Chilan Berry. Default is FALSE.}

\item{random}{Character string. If "yes", a random uniformly distributed integer between 85 and 100 is used. If "minmax", the min and max values are returned. If anything else, the average is used. Default is "yes".}
}
\value{
The calculated damage.
}
\description{
This function calculates the damage inflicted during a Pokemon battle.
It uses the formula from Generation IV.
}
\note{
this function does not handle the specific conditions mentioned (like Reflect, Light Screen, Explosion, Selfdestruct, etc.).
You would need to handle these conditions before passing the values to this function
}
