% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random.R
\name{randomPokemon}
\alias{randomPokemon}
\title{Returns a random Pokemon}
\usage{
randomPokemon(n = 1, ..., p = English, replace = FALSE)
}
\arguments{
\item{n}{How many}

\item{p}{What column to return (Ndex, English, form, form2, type, type2 can be used)}

\item{replace}{Should sampling be with replacement?}

\item{type}{Filter on type}

\item{type2}{Filter on type2}

\item{swap}{Allow type to match type2 and vicea-versa}

\item{data}{Override the data (Must have type and type2 (If type 2 is not ANY or swap is TRUE))}
}
\description{
Returns a random Pokemon
}
\examples{
# Get a random Pokemon
randomPokemon()
# > [1] "Koraidon"

# Get a random Dragon type
randomPokemon(type = "Dragon")
> [1] "Zekrom"

# Get 6 random Dragon Pokemon
randomPokemon(6, type = "Dragon")
# > [1] "Dragapult" "Jangmo-o" "Rayquaza" "Zekrom" "Garchomp" "Haxorus"
# Get a random Dragon-Electric Pokemon
randomPokemon(type = "Dragon", type2 = "Electric")
# > [1] "Zekrom"

# Get a random Dragon-Electric or Electric-Dragon Pokemon
randomPokemon(type = "Dragon", type2 = "Electric", swap = TRUE)
# > [1] "Miraidon"

# Get All random Dragon-\* or \*-Dragon Pokemon
randomPokemon(n = -1, type = "Dragon", swap = TRUE)
# > [1] "Zekrom" "Dreepy" "Guzzlord" "Archaludon" "Dragonite" "Raging Bolt" "Frigibax" "Zweilous" "Goodra"
# > [10] "Rayquaza" "Regidrago" "Garchomp" "Kingdra" "Roaring Moon" "Drakloak" "Jangmo-o" "Bagon" "Eternatus"
# > [19] "Turtonator" "Altaria" "Goomy" "Baxcalibur" "Hydreigon" "Druddigon" "Hydrapple" "Miraidon" "Dragapult"
# > [28] "Salamence" "Drampa" "Fraxure" "Duraludon" "Zygarde" "Cyclizar" "Haxorus" "Dipplin" "Dratini"
# > [37] "Dracovish" "Noibat" "Gouging Fire" "Latios" "Kyurem" "Latias" "Palkia" "Walking Wake" "Kommo-o"
# > [46] "Sliggoo" "Vibrava" "Giratina" "Dracozolt" "Flygon" "Tyrunt" "Shelgon" "Tatsugiri" "Dragalge"
# > [55] "Axew" "Zygarde50" "Arctibax" "Tyrantrum" "Flapple" "Gible" "Deino" "Applin" "Exeggutor"
# > [64] "Hakamo-o" "Noivern" "Koraidon" "Dragonair" "Dialga" "Gabite" "Appletun" "Reshiram" "Naganadel"
# > [73] "Zygarde10"

# Return the dex number instead
randomPokemon(n = -1, type = "Dragon", swap = TRUE, p = Ndex)
# > [1] "#1011" "#1019" "#0371" "#0633" "#0634" "#0841" "#0334" "#0691" "#0644" "#0718" "#0782" "#0381" "#0445" "#0804" "#0840" "#0704" "#0799"
# > [18] "#0886" "#0380" "#0330" "#0842" "#0880" "#0714" "#0147" "#0895" "#0646" "#0635" "#0705" "#0884" "#1018" "#0967" "#0384" "#0372" "#0373"
# > [35] "#0621" "#0780" "#0696" "#0887" "#0483" "#1008" "#0882" "#0706" "#0697" "#1005" "#0784" "#0610" "#0978" "#0998" "#0890" "#0997" "#0148"
# > [52] "#0443" "#0715" "#0776" "#0230" "#0611" "#0783" "#0996" "#1020" "#0103" "#0643" "#1009" "#0487" "#0329" "#0885" "#0612" "#0444" "#0149"
# > [69] "#1021" "#1007" "#0484"

}
