% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{read_data}
\alias{read_data}
\title{Read Data from CSV or RDA}
\usage{
read_data(data, root, ns = asNamespace("ZekDex"), one = TRUE, g = NULL)
}
\arguments{
\item{data}{The name of the data to load.}

\item{root}{The root directory where the file is located.}

\item{ns}{The namespace from which to load the data.}

\item{one}{Is there one item in the rda file?}
}
\description{
This function reads data from a CSV or RDA file, or from a package namespace.
}
