% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{recursive_wrap}
\alias{recursive_wrap}
\title{Vectorized version of the .recursive_wrap function}
\usage{
recursive_wrap(str, .open = "(", .close = ")?")
}
\arguments{
\item{str}{A vector of strings to be wrapped}
}
\value{
A vector of strings that are recursively wrapped with the specified opening and closing characters
}
\description{
Vectorized version of the .recursive_wrap function
}
