% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regionalDex}
\alias{regionalDex}
\title{Pokemon Dataset}
\format{
A data frame with the following variables:
\describe{
\item{ndex}{National Pokedex number.}
\item{name}{Name of the Pokemon.}
\item{form}{Form of the Pokemon.}
\item{regional}{Regional form of the Pokemon.}
\item{type}{Primary type of the Pokemon.}
\item{type2}{Secondary type of the Pokemon.}
\item{generation}{Generation of the Pokemon.}
\item{isLegendary}{Boolean indicating if the Pokemon is Legendary.}
\item{isMythical}{Boolean indicating if the Pokemon is Mythical.}
\item{isUltraBeast}{Boolean indicating if the Pokemon is an Ultra Beast.}
\item{family}{Family of the Pokemon.}
\item{size}{Size of the Pokemon.}
\item{NA_Kanto_Kdex}{Kanto Pokedex number.}
\item{NA_Johto_Jdex}{Johto Pokedex number.}
\item{GenerationIII_Hoenn_Hdex}{Hoenn Pokedex number in Generation III.}
\item{GenerationVI_Hoenn_Hdex}{Hoenn Pokedex number in Generation VI.}
\item{NA_Sinnoh_Sdex}{Sinnoh Pokedex number.}
\item{BlackandWhite_Unova_Udex}{Unova Pokedex number in Black and White.}
\item{Black2andWhite2_Unova_Udex}{Unova Pokedex number in Black 2 and White 2.}
\item{NA_Blueberry_BBdex}{Blueberry Pokedex number.}
\item{NA_Kalos_CeKdex}{Central Kalos Pokedex number.}
\item{NA_Kalos_CoKdex}{Coastal Kalos Pokedex number.}
\item{NA_Kalos_MoKdex}{Mountain Kalos Pokedex number.}
\item{SunandMoon_Alola_Adex}{Alola Pokedex number in Sun and Moon.}
\item{SunandMoon_Alola_Melemele}{Melemele Pokedex number in Sun and Moon.}
\item{SunandMoon_Alola_Akala}{Akala Pokedex number in Sun and Moon.}
\item{SunandMoon_Alola_Ula'ula}{Ula'ula Pokedex number in Sun and Moon.}
\item{SunandMoon_Alola_Poni}{Poni Pokedex number in Sun and Moon.}
\item{UltraSunandUltraMoon_Alola_Adex}{Alola Pokedex number in Ultra Sun and Ultra Moon.}
\item{UltraSunandUltraMoon_Alola_Melemele}{Melemele Pokedex number in Ultra Sun and Ultra Moon.}
\item{UltraSunandUltraMoon_Alola_Akala}{Akala Pokedex number in Ultra Sun and Ultra Moon.}
\item{UltraSunandUltraMoon_Alola_Ula'ula}{Ula'ula Pokedex number in Ultra Sun and Ultra Moon.}
\item{UltraSunandUltraMoon_Alola_Poni}{Poni Pokedex number in Ultra Sun and Ultra Moon.}
\item{NA_Galar_Gdex}{Galar Pokedex number.}
\item{NA_IsleofArmor_IoAdex}{Isle of Armor Pokedex number.}
\item{NA_CrownTundra_CTdex}{Crown Tundra Pokedex number.}
\item{NA_Hisui_LAdex}{Hisui Pokedex number.}
\item{NA_Paldea_Pdex}{Paldea Pokedex number.}
\item{NA_Kitakami_Kdex}{Kitakami Pokedex number.}
}
}
\usage{
regionalDex
}
\description{
This dataset contains information about various Pokemon.
}
\keyword{datasets}
