% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats}
\alias{stats}
\title{Pokemon Data}
\format{
A data frame with 91 rows and 11 variables:
\describe{
\item{ndex}{\code{integer} National Pokédex number}
\item{name}{\code{character} Pokémon name}
\item{HP}{\code{integer} Hit Points}
\item{Attack}{\code{integer} Attack strength}
\item{Defense}{\code{integer} Defensive strength}
\item{Speed}{\code{integer} Speed}
\item{Special}{\code{integer} Special attack or special defense strength}
\item{Total}{\code{integer} Total of all stats}
\item{Average}{\code{numeric} Average of all stats}
\item{Gen}{\code{factor} Generation}
\item{SpAttack}{\code{integer} Special Attack strength (NA for Gen I)}
\item{SpDefense}{\code{integer} Special Defense strength (NA for Gen I)}
}
}
\source{
Bulbapedia
}
\usage{
stats
}
\description{
Base stats of Pokémon including various forms / sizes of Pokémon.
}
\keyword{datasets}
