% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typeChartWide}
\alias{typeChartWide}
\title{Pokémon Type Effectiveness Dataset}
\format{
A data frame with 19 variables (1 for the attacking type and 18 for each defending type):
\describe{
\item{Attacking}{\code{factor} The type of the attacking Pokémon's move. There are 18 types in total.}
\item{Defending_\emph{}{\code{numeric} The effectiveness of the attacking type against the defending type. This is represented by a multiplier (0, 0.5, 1, 2) applied to the attack's damage. Each 'Defending_}' represents a different Pokémon type.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/Type}{Bulbapedia Type}
}
\usage{
typeChartWide
}
\description{
This dataset provides a matrix of effectiveness multipliers for different types of Pokémon attacks.
Each Pokémon has a type, such as Fire, Water, Grass, etc., and each type has different effectiveness against other types.
This effectiveness is represented by a multiplier that is applied to the attack's damage.
}
\details{
The dataset is structured as a wide format, with each column representing a defending Pokémon's type and
each row representing an attacking Pokémon's type.
}
\examples{
data(typeChartWide)
}
\keyword{datasets}
