% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{types}
\alias{types}
\title{Pokemon Types Dataset}
\format{
A data frame with one column:
\describe{
\item{types}{\code{character} The types of Pokemon. Possible types include "Grass", "Fire", "Water", "Bug", "Normal", "Dark", "Poison", "Electric", "Ground", "Ice", "Fairy", "Steel", "Fighting", "Psychic", "Rock", "Ghost", "Dragon", "Flying".}
}
}
\source{
Bulbapedia
}
\usage{
types
}
\description{
This dataset contains information about various Pokemon types.
}
\keyword{datasets}
