% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight}
\alias{weight}
\title{Pokemon Weight Dataset}
\format{
A data frame with 8 variables:
\describe{
\item{name}{\code{character} The name of the Pokémon.}
\item{form}{\code{character} The form of the Pokémon. This refers to the different physical forms that Pokémon can take. 'NA' if the Pokémon does not have another form.}
\item{lbs}{\code{numeric} The weight of the Pokémon in pounds.}
\item{kg}{\code{numeric} The weight of the Pokémon in kilograms.}
\item{lightest}{\code{integer} The rank of the Pokémon from lightest (1) to heaviest (1181).}
\item{heaviest}{\code{integer} The rank of the Pokémon from heaviest (1) to lightest (1181).}
\item{regional}{\code{character} The regional form of the Pokémon. This refers to the variations of Pokémon that appear differently and have different attributes compared to their counterparts in other regions. 'NA' if the Pokémon does not have a regional form.}
\item{MegaOrPrimal}{\code{character} Indicates if the Pokémon is in a Mega or Primal form. These are special forms that certain Pokémon can take.}
}
}
\source{
\href{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_weight}{Bulbapedia Weight}
}
\usage{
weight
}
\description{
This dataset contains information about the weight of various Pokémon.
Each Pokémon's weight is recorded in both pounds (lbs) and kilograms (kg), providing flexibility for different uses.
The dataset also includes details about the lightest and heaviest Pokémon, their regional forms, and whether they have a Mega or Primal form.
This can be useful for a variety of analyses, such as studying the distribution of Pokémon weights,
identifying trends and patterns, or informing game strategies.
}
\examples{
data(weight)
}
\keyword{datasets}
